Landmark MDS

A fast dimensionality reduction method scaleable to large numbers of samples. Landmark Multi-Dimensional Scaling (LMDS) is an extension of classical 'Torgerson MDS', but rather than calculating a complete distance matrix between all pairs of samples, only the distances between a set of landmarks and the samples are calculated.

Usage

lmds(
  x,
  ndim = 3,
  distance_method = c("euclidean", "pearson", "spearman", "cosine", "chisquared",
    "hamming", "kullback", "manhattan", "maximum", "canberra", "minkowski"),
  landmark_method = c("sample"),
  num_landmarks = 500
)

Arguments

x

A matrix, optionally sparse.

ndim

The number of dimensions

distance_method

The distance metric to use. Options are "euclidean" (default), "pearson", "spearman", "cosine", "manhattan".

landmark_method

The landmark selection method to use. Options are "sample" (default).

num_landmarks

The number of landmarks to use,

Value

The dimensionality reduction in the form of a nrow(x) by ndim matrix.

Examples

library(Matrix)
x <- Matrix::rsparsematrix(1000, 1000, .01)
lmds(x, ndim = 3)
#>              comp_1        comp_2        comp_3
#>    [1,] -0.32744001 -5.469474e-03  0.2011463844
#>    [2,] -0.30961065  7.276987e-03  0.1758819494
#>    [3,] -0.27842870 -1.524229e-02  0.1439829652
#>    [4,] -0.27421847  1.769585e-02  0.1472096580
#>    [5,] -0.29360131  3.106276e-02  0.1520043726
#>    [6,] -0.29997098  8.199960e-03  0.1707871533
#>    [7,] -0.26624939  1.232338e-01  0.0711097350
#>    [8,] -0.30737189  1.041641e-02  0.1771724597
#>    [9,] -0.22493806  2.232000e-02  0.1693231072
#>   [10,] -0.31778718  5.255387e-03  0.2107662314
#>   [11,] -0.31039231  2.373425e-03  0.1803610160
#>   [12,] -0.29917712  4.012345e-02  0.1544087496
#>   [13,] -0.31523567  1.220795e-03  0.1852644624
#>   [14,] -0.31168122  1.228987e-03  0.1837770268
#>   [15,] -0.30141395 -2.732933e-03  0.1693055511
#>   [16,] -0.29364581  7.102768e-03  0.1707395530
#>   [17,] -0.30041224  8.501822e-03  0.1653830211
#>   [18,] -0.25003904  1.117788e-01 -0.0065637258
#>   [19,]  0.10449888  2.395394e-02  0.1579795778
#>   [20,] -0.29388494 -2.836360e-03  0.1733198862
#>   [21,] -0.30625486 -3.111152e-03  0.1801069618
#>   [22,] -0.31735536 -1.002496e-02  0.1873508987
#>   [23,] -0.28942500  2.159636e-02  0.1785463856
#>   [24,] -0.33088645  6.075026e-03  0.2117371341
#>   [25,] -0.31166236  2.426848e-03  0.1901467240
#>   [26,] -0.28799897 -1.730064e-03  0.1507587341
#>   [27,] -0.28426010  2.280831e-02  0.1691339591
#>   [28,] -0.29619910  8.898059e-03  0.1706198641
#>   [29,] -0.31005313  5.675196e-03  0.1703021425
#>   [30,] -0.27494324  9.098009e-02  0.1388913098
#>   [31,] -0.30690810 -2.208061e-03  0.1843508494
#>   [32,] -0.30497289  3.694296e-04  0.1185371769
#>   [33,] -0.29816431  5.044170e-03  0.1761221320
#>   [34,] -0.28480684  1.445586e-02  0.1698831275
#>   [35,] -0.32617775 -5.236507e-03  0.1935746672
#>   [36,] -0.28808314  1.393241e-02  0.1502638178
#>   [37,] -0.33484667 -1.665885e-02  0.2072908658
#>   [38,] -0.31578009  8.201405e-04  0.1902414923
#>   [39,] -0.31006991 -3.326576e-02  0.2003384272
#>   [40,] -0.32603517 -8.824005e-03  0.1973407610
#>   [41,] -0.29496099  1.014383e-02  0.1637747472
#>   [42,] -0.31787139  4.704688e-03  0.1831738388
#>   [43,] -0.29797343  9.741216e-03  0.1740012049
#>   [44,] -0.27316793  2.217718e-02  0.1824985788
#>   [45,] -0.31866922 -6.535551e-03  0.1913566708
#>   [46,] -0.29314545  1.391345e-02  0.1240611882
#>   [47,] -0.28459325  1.333933e-02  0.1573623806
#>   [48,] -0.29351449  6.957326e-02  0.1626723798
#>   [49,] -0.30246758  1.300491e-02  0.1763916235
#>   [50,] -0.28682948  4.831052e-03  0.1831493773
#>   [51,] -0.32129265 -4.170638e-03  0.1955561151
#>   [52,] -0.28665541  7.672864e-03  0.1547183711
#>   [53,] -0.28670509  1.724769e-02  0.1500867609
#>   [54,] -0.26141791 -1.608047e-02  0.1478693734
#>   [55,] -0.29998693  9.073688e-03  0.1668136604
#>   [56,] -0.31803207  1.935029e-03  0.1846124629
#>   [57,] -0.29941630  7.388136e-04  0.1720216627
#>   [58,] -0.32317259 -5.651440e-03  0.1957196781
#>   [59,] -0.29570033  5.489351e-03  0.1708802206
#>   [60,] -0.26577293  2.624833e-02  0.1383043569
#>   [61,] -0.32401615 -5.080582e-03  0.1971816833
#>   [62,] -0.31093213 -1.747865e-03  0.1829176016
#>   [63,] -0.28319298  1.121859e-02  0.1575471156
#>   [64,] -0.29553005  1.147207e-02  0.1512745574
#>   [65,] -0.28799753  1.261739e-02  0.1651537422
#>   [66,] -0.30489975  7.743198e-03  0.1366783367
#>   [67,] -0.33168503 -8.177589e-03  0.1998974273
#>   [68,] -0.29429940 -1.735309e-03  0.1567105500
#>   [69,] -0.29412352  2.165458e-02  0.1907880642
#>   [70,] -0.28561217  1.890959e-02  0.1621928106
#>   [71,] -0.29631448  5.617999e-02  0.2120682274
#>   [72,] -0.29323762  3.537046e-03  0.1879630680
#>   [73,] -0.23454667  4.764425e-02  0.1633647526
#>   [74,] -0.30632212  6.479673e-04  0.1769540010
#>   [75,] -0.30391586  2.249382e-03  0.1783979391
#>   [76,] -0.28784982  1.813838e-02  0.1644924679
#>   [77,] -0.29011278  3.120805e-02  0.1574900320
#>   [78,] -0.30968787  6.351369e-04  0.1773067265
#>   [79,] -0.29937378  1.519606e-02  0.1677859682
#>   [80,] -0.31021390 -3.100171e-03  0.1831465761
#>   [81,] -0.26219118 -2.282603e-02  0.1680917323
#>   [82,] -0.29759112  1.851609e-02  0.1640759281
#>   [83,] -0.30316382  2.242029e-02  0.1693853870
#>   [84,] -0.33805950 -1.416640e-02  0.2047671599
#>   [85,] -0.28368005  2.019595e-02  0.1666320527
#>   [86,] -0.29011395  2.044520e-02  0.1743345872
#>   [87,] -0.30860659  1.408926e-02  0.1958547143
#>   [88,] -0.28334746  1.401979e-02  0.1657097366
#>   [89,] -0.32103577 -3.357036e-03  0.1957087567
#>   [90,] -0.28815040  3.096339e-02  0.1709456556
#>   [91,] -0.32576670 -5.098756e-03  0.1892261278
#>   [92,] -0.29132964 -2.565836e-03  0.1743604495
#>   [93,] -0.32982221 -6.556476e-03  0.1975352725
#>   [94,] -0.30554883  3.894985e-03  0.1397934613
#>   [95,] -0.24838017 -2.349584e-03  0.1474757843
#>   [96,] -0.31401268 -1.900981e-03  0.1811242292
#>   [97,] -0.31732298 -1.517773e-03  0.1963866422
#>   [98,] -0.23948416  1.466289e-02  0.1449117041
#>   [99,] -0.23137056  4.605785e-02  0.1326246330
#>  [100,] -0.27468866  2.097006e-02  0.1473358065
#>  [101,] -0.29653305  1.179780e-02  0.1578013992
#>  [102,] -0.29944805  1.114192e-02  0.1788411569
#>  [103,] -0.27444158  1.732467e-02  0.1474853167
#>  [104,] -0.29364359  9.803805e-03  0.1604971976
#>  [105,] -0.29714605 -5.738810e-03  0.1663124750
#>  [106,] -0.27468825  2.527417e-02  0.1414082233
#>  [107,] -0.33024543 -8.220095e-03  0.1984179751
#>  [108,] -0.28740739  6.174416e-02  0.1662962400
#>  [109,] -0.29763315  1.718350e-02  0.1592366025
#>  [110,] -0.28485012  1.561450e-03  0.1538915614
#>  [111,] -0.32445450 -3.114352e-03  0.1946203979
#>  [112,] -0.26985451  1.350888e-02  0.1698692703
#>  [113,] -0.29249380  4.383301e-02  0.1811034683
#>  [114,] -0.30143629  4.050977e-03  0.1881416653
#>  [115,] -0.31997311 -2.817364e-03  0.1855985792
#>  [116,] -0.27568134  1.709223e-02  0.1518263139
#>  [117,] -0.27519845  2.104546e-02  0.1554841996
#>  [118,] -0.31156055  1.129601e-03  0.1669764328
#>  [119,] -0.25501150  3.177933e-02  0.1370453212
#>  [120,] -0.31892504 -8.017074e-03  0.1858541165
#>  [121,] -0.30920677 -2.679906e-03  0.1875281927
#>  [122,] -0.27362318  2.740109e-02  0.0524599794
#>  [123,] -0.30065606  1.260535e-02  0.1365687346
#>  [124,] -0.28834278  3.350159e-02  0.1757930606
#>  [125,] -0.31658740 -6.724731e-03  0.1959261571
#>  [126,] -0.32747195 -3.238430e-03  0.1953441171
#>  [127,] -0.28274793  1.213484e-01  0.1511303104
#>  [128,] -0.29003628  1.162572e-02  0.1626393062
#>  [129,] -0.29802648  7.477207e-03  0.1659454850
#>  [130,] -0.30441091 -4.828176e-03  0.1782577391
#>  [131,] -0.27543077  3.264833e-02  0.1536288547
#>  [132,] -0.26616291  2.078352e-01  0.0346279267
#>  [133,] -0.30559921  5.419938e-03  0.1779280031
#>  [134,] -0.27690605  1.051629e-02  0.1514965659
#>  [135,] -0.28806672  6.305463e-03  0.1413567667
#>  [136,] -0.27916414  3.999972e-03  0.1693850095
#>  [137,] -0.27819372 -1.248077e-02  0.1739272977
#>  [138,] -0.30428414  3.879677e-03  0.1825783726
#>  [139,] -0.27339389  6.097941e-02  0.1769781333
#>  [140,] -0.25155163 -8.385647e-02  0.1998650020
#>  [141,] -0.30245389  2.748000e-02  0.1431084795
#>  [142,] -0.27777524  2.456723e-02  0.1531159784
#>  [143,] -0.30237822  4.921276e-03  0.1775355452
#>  [144,] -0.28824231  2.284836e-02  0.1577635967
#>  [145,] -0.26436426  1.064019e-02  0.1493912370
#>  [146,] -0.28090334  8.306763e-03  0.1655796311
#>  [147,] -0.29664232 -2.454239e-02  0.1821911639
#>  [148,] -0.29966545 -7.183526e-03  0.2003795197
#>  [149,] -0.30451550  1.538938e-03  0.1777190556
#>  [150,] -0.29250273  1.642942e-02  0.1892581843
#>  [151,] -0.29965067 -7.148490e-03  0.1516159336
#>  [152,] -0.32528873  6.353066e-03  0.1813249977
#>  [153,] -0.28621352  1.315616e-02  0.1672255867
#>  [154,] -0.30517491  1.046037e-02  0.1797472339
#>  [155,] -0.28969209 -1.171107e-01  0.0739529558
#>  [156,] -0.30308482  2.737372e-02  0.1565489370
#>  [157,] -0.31553939  1.384118e-03  0.1875330929
#>  [158,] -0.29237562  3.335897e-02  0.1730810095
#>  [159,] -0.31097908  4.207234e-04  0.1784809458
#>  [160,] -0.30971121  9.971684e-03  0.1715864066
#>  [161,] -0.29414397  9.544357e-03  0.1595919285
#>  [162,] -0.29990804  4.428986e-03  0.1750939319
#>  [163,] -0.28546920  1.335748e-01  0.1281949355
#>  [164,] -0.30230354  5.234302e-03  0.1818888399
#>  [165,] -0.29035625 -6.768608e-02  0.2081548838
#>  [166,] -0.30756975  9.973663e-03  0.1852107646
#>  [167,] -0.31317064 -2.366169e-02  0.1900129277
#>  [168,] -0.30307377  1.297496e-02  0.1649133526
#>  [169,] -0.27052230  1.854736e-02  0.1532106219
#>  [170,] -0.26003623  9.651962e-03  0.1613192738
#>  [171,] -0.28575803 -1.955368e-03  0.1660930875
#>  [172,] -0.28249529  2.254136e-02  0.1496724096
#>  [173,] -0.28848994  3.486774e-03  0.1610368633
#>  [174,] -0.29541090  5.577731e-03  0.1670100824
#>  [175,] -0.29811128  1.025738e-02  0.1676710149
#>  [176,] -0.31264335  2.354888e-04  0.1819081543
#>  [177,] -0.26822751  4.121542e-02  0.1538165400
#>  [178,] -0.29477546  1.169059e-02  0.1670564000
#>  [179,] -0.28211429  5.459932e-03  0.1102751073
#>  [180,] -0.28752445  3.789080e-03  0.1861001325
#>  [181,] -0.32231275 -2.093544e-03  0.1835395627
#>  [182,] -0.30158554 -2.974679e-02  0.1939722608
#>  [183,] -0.30150166  8.482574e-03  0.1669660749
#>  [184,] -0.28754167  2.149252e-02  0.2024403848
#>  [185,] -0.29282379  1.646412e-02  0.1775679864
#>  [186,] -0.29006085  2.448593e-01  0.3444840081
#>  [187,] -0.29800938  1.669085e-02  0.1787930038
#>  [188,] -0.31252173 -5.085301e-03  0.1876522606
#>  [189,] -0.27758347  1.694962e-02  0.1504770495
#>  [190,] -0.26380453  1.151360e-02  0.1315697120
#>  [191,] -0.27659718  1.837037e-02  0.1518003622
#>  [192,] -0.30808109  6.387319e-03  0.1597883682
#>  [193,] -0.28806277  2.590880e-03  0.2860100913
#>  [194,] -0.32093290  5.209357e-04  0.1879602733
#>  [195,] -0.31267602 -2.857873e-02  0.1986797968
#>  [196,] -0.25889762  4.304835e-01 -0.0928237157
#>  [197,] -0.32086548 -3.763268e-04  0.1813799439
#>  [198,] -0.28390357 -5.051504e-03  0.1649796854
#>  [199,] -0.28381089  8.576784e-03  0.1688408043
#>  [200,] -0.28989477  1.592683e-03  0.1618350693
#>  [201,] -0.30950314  5.883482e-02  0.1567799294
#>  [202,] -0.32120649 -4.451050e-03  0.1925518320
#>  [203,] -0.34109027 -8.452883e-03  0.1490167051
#>  [204,] -0.30886904  8.349366e-03  0.1780963850
#>  [205,] -0.27791299  9.690702e-03  0.1691147332
#>  [206,] -0.28417755  5.966258e-02  0.1397398709
#>  [207,] -0.28128123  1.856025e-02  0.1456587050
#>  [208,] -0.29484268  7.113150e-03  0.2113990275
#>  [209,] -0.31266887  5.679037e-03  0.1814758523
#>  [210,] -0.32148017 -4.320823e-03  0.1866897109
#>  [211,] -0.24381705 -6.787615e-03  0.1973465155
#>  [212,] -0.27768425  8.560123e-02  0.1689907553
#>  [213,] -0.31716302 -5.980208e-03  0.1835889590
#>  [214,] -0.30168312  8.556202e-03  0.1882130688
#>  [215,] -0.27786342  1.910357e-02  0.1499916250
#>  [216,] -0.31107573 -4.183660e-03  0.1794746023
#>  [217,] -0.28900583 -7.945350e-03  0.1661368441
#>  [218,]  0.50000000  5.775870e-02  0.1599968797
#>  [219,] -0.31004562  7.445835e-04  0.1684915815
#>  [220,] -0.30749850  1.009546e-02  0.2129597403
#>  [221,] -0.29755924 -1.035287e-03  0.1620819609
#>  [222,] -0.27697191  1.001658e-03  0.1528529322
#>  [223,] -0.31910878 -1.781282e-02  0.1528647316
#>  [224,] -0.30617235  1.023483e-02  0.1731168725
#>  [225,] -0.31624469 -4.995137e-03  0.2025524718
#>  [226,] -0.21459260  2.594372e-02  0.1896958069
#>  [227,] -0.29524820  5.812117e-03  0.1621202942
#>  [228,] -0.30064770  1.437385e-02  0.1543881555
#>  [229,] -0.30852331  3.851383e-03  0.1790116020
#>  [230,] -0.29956932  1.852129e-03  0.1620295006
#>  [231,] -0.30853524  5.340494e-03  0.1696646208
#>  [232,] -0.28146678 -2.892787e-02  0.1600866343
#>  [233,] -0.27698154  5.819981e-02  0.1288188012
#>  [234,] -0.30764433  5.616240e-03  0.1681267031
#>  [235,] -0.26831099  2.915904e-02  0.1369023691
#>  [236,] -0.29342894  1.001331e-02  0.1818769773
#>  [237,] -0.28402003  1.420625e-02  0.1688949704
#>  [238,] -0.30152195  2.979941e-03  0.1696098580
#>  [239,] -0.29368674  2.350922e-02  0.1736118578
#>  [240,] -0.28050188  1.623926e-02  0.1602156842
#>  [241,] -0.30132334  2.422813e-03  0.1638693240
#>  [242,] -0.32230402 -2.343686e-03  0.1966602881
#>  [243,] -0.32814613 -9.799805e-03  0.1980092273
#>  [244,] -0.28890989  9.740870e-03  0.1565252292
#>  [245,] -0.28066536  2.950059e-03  0.1510563271
#>  [246,] -0.32240043 -4.329130e-04  0.1912730662
#>  [247,] -0.29470287 -4.458896e-03  0.1870475625
#>  [248,] -0.28207023  6.726066e-03  0.1242229329
#>  [249,] -0.28758764  7.673789e-03  0.1609687760
#>  [250,] -0.30643755  4.881310e-03  0.1852321761
#>  [251,] -0.33495034 -1.357272e-02  0.2012170526
#>  [252,] -0.28520046  2.148536e-02  0.1686046344
#>  [253,] -0.27485996 -4.387539e-02  0.0897395380
#>  [254,] -0.28250601  7.467786e-03  0.1572992000
#>  [255,] -0.28926418 -6.974284e-04  0.1672700041
#>  [256,] -0.29351781  1.153690e-02  0.1644533775
#>  [257,] -0.25555324  9.506886e-03  0.1533699537
#>  [258,] -0.28769550  2.752194e-02  0.1373042180
#>  [259,] -0.30052901  3.928777e-02  0.1820125573
#>  [260,] -0.28880988  1.608892e-02  0.1603051209
#>  [261,] -0.28437877  4.008870e-02  0.1531044939
#>  [262,] -0.29949303  4.477339e-03  0.1624665591
#>  [263,] -0.29970317  8.862897e-04  0.1923032883
#>  [264,] -0.29047805  1.834707e-02  0.1471050018
#>  [265,] -0.30193397  2.252633e-02  0.1475140181
#>  [266,] -0.31067733 -1.499670e-03  0.1782233044
#>  [267,] -0.30746447 -5.137301e-03  0.1693992074
#>  [268,] -0.28535397  5.275250e-03  0.1369164409
#>  [269,] -0.29710548  1.024253e-02  0.1814520903
#>  [270,] -0.30864906  2.717552e-02  0.1673819746
#>  [271,] -0.29565138  1.342781e-02  0.1694769873
#>  [272,] -0.30402556  8.947438e-03  0.1695379871
#>  [273,] -0.26786850  2.749875e-02  0.1768274021
#>  [274,] -0.31805858  1.204985e-03  0.1979138425
#>  [275,] -0.28190241  1.873516e-02  0.1510418936
#>  [276,] -0.30726665 -2.666483e-03  0.1815707082
#>  [277,] -0.29825729  2.738474e-03  0.1531075601
#>  [278,] -0.32208518  1.218682e-03  0.1570896726
#>  [279,] -0.28985140  1.794732e-02  0.1581751366
#>  [280,] -0.28404145  1.462044e-02  0.1605506777
#>  [281,] -0.28050831  1.592956e-02  0.1497548342
#>  [282,] -0.30073883  1.586933e-02  0.1669199133
#>  [283,] -0.27888113  2.049302e-02  0.1528607206
#>  [284,] -0.23275051 -6.474762e-02  0.1844015490
#>  [285,] -0.29312735  4.986452e-04  0.1890491101
#>  [286,] -0.29339878 -5.958080e-03  0.1778622524
#>  [287,] -0.28665439  1.138895e-01  0.1433585133
#>  [288,] -0.27839287  1.891027e-02  0.1503652715
#>  [289,] -0.34009773  5.480894e-03  0.1629197771
#>  [290,] -0.28050245  8.592271e-03  0.1612425352
#>  [291,] -0.29406379  3.951073e-03  0.1690069021
#>  [292,] -0.29996184  2.741221e-02  0.1573440654
#>  [293,] -0.30413468  1.166334e-02  0.1835655399
#>  [294,] -0.28264117 -2.587579e-03  0.1796413966
#>  [295,] -0.29617738  4.542848e-02  0.1929882656
#>  [296,] -0.31654665  1.479708e-02  0.1910908922
#>  [297,] -0.29146400  9.512871e-03  0.1683354378
#>  [298,] -0.29751391  1.051980e-02  0.1603707407
#>  [299,] -0.30810529  4.165838e-03  0.1772799750
#>  [300,] -0.27757052 -1.218046e-02  0.1391598806
#>  [301,] -0.29529216  1.326133e-02  0.1637151983
#>  [302,] -0.32649752 -1.130035e-02  0.1872923894
#>  [303,] -0.29498017  3.857165e-03  0.1661338598
#>  [304,] -0.31317061 -1.469397e-03  0.1846505160
#>  [305,] -0.29668790  1.149689e-02  0.1843035868
#>  [306,] -0.26840693  2.101674e-02  0.1490103927
#>  [307,] -0.30180572  1.139007e-02  0.1672313919
#>  [308,] -0.30925909  1.848252e-02  0.1843505608
#>  [309,] -0.28313892  7.961409e-03  0.1240873118
#>  [310,] -0.29171068  5.838023e-03  0.1668673964
#>  [311,] -0.30337643  3.103600e-03  0.1728573542
#>  [312,] -0.32592874 -6.282417e-03  0.1939167373
#>  [313,] -0.29976530  9.153145e-03  0.1690593143
#>  [314,] -0.30175015  7.483065e-03  0.1699502932
#>  [315,] -0.28575150  3.193589e-03  0.1685717208
#>  [316,] -0.29163152  1.198057e-02  0.1855784321
#>  [317,] -0.27618821  7.741375e-02  0.1089009242
#>  [318,] -0.30594932  2.777624e-02  0.1558420855
#>  [319,] -0.26774977  2.995933e-02  0.1418985179
#>  [320,] -0.31263462 -3.366273e-03  0.2032346039
#>  [321,] -0.31354530 -8.166018e-03  0.1811001459
#>  [322,] -0.32505063 -3.519139e-03  0.1983899650
#>  [323,] -0.28761212  1.787048e-02  0.1627051362
#>  [324,] -0.28706072  1.496618e-02  0.1620416942
#>  [325,] -0.28657422 -5.566640e-03  0.1700799074
#>  [326,] -0.28563391  5.514978e-03  0.1627262303
#>  [327,] -0.27925608  1.341540e-02  0.1583551348
#>  [328,] -0.31359512 -8.718635e-03  0.1972554029
#>  [329,] -0.27944715  8.600287e-03  0.1623240257
#>  [330,] -0.23092515  3.912402e-02  0.1365207853
#>  [331,] -0.31094108 -2.746913e-04  0.1771338194
#>  [332,] -0.29027291  1.714703e-02  0.1593940129
#>  [333,] -0.32348163  6.068169e-02  0.2516229764
#>  [334,] -0.31539834  1.770418e-03  0.1780966182
#>  [335,] -0.29915610  2.987967e-02  0.1651686605
#>  [336,] -0.30440365 -4.185872e-04  0.1821040835
#>  [337,] -0.31245371 -2.178244e-03  0.1900383206
#>  [338,] -0.28893391 -2.246207e-04  0.1444018391
#>  [339,] -0.31561184 -2.241327e-02  0.1758447909
#>  [340,] -0.27139745  2.004873e-02  0.1529022600
#>  [341,] -0.31930037  2.490921e-03  0.1860223108
#>  [342,] -0.27985964  4.107594e-02  0.1436653163
#>  [343,] -0.27257704  9.921629e-03  0.1538956159
#>  [344,] -0.30187820  1.365402e-02  0.1691117434
#>  [345,] -0.29025126  1.158638e-02  0.1590863562
#>  [346,] -0.29112116  1.905636e-02  0.1606951056
#>  [347,] -0.31360726 -2.635791e-03  0.1754817029
#>  [348,] -0.28316852  3.937398e-02  0.1508349322
#>  [349,] -0.26123908  1.334974e-02  0.1557585119
#>  [350,] -0.27578368  2.829860e-02  0.1629137502
#>  [351,] -0.38744426 -2.227456e-02  0.1528551019
#>  [352,] -0.26064746  1.643576e-02  0.1525690043
#>  [353,] -0.25566820  3.519002e-02  0.1348172041
#>  [354,] -0.28816709 -2.031258e-02  0.1611919120
#>  [355,] -0.27808870  1.156540e-01  0.1194301736
#>  [356,] -0.27943954  1.727154e-02  0.1355392057
#>  [357,] -0.31674556 -5.197551e-03  0.1799767249
#>  [358,] -0.26917922 -5.310189e-02  0.0802387014
#>  [359,] -0.27407440  3.326690e-02  0.1646091717
#>  [360,] -0.30571433  1.356054e-02  0.1715224666
#>  [361,] -0.28589055  1.458540e-02  0.1581204296
#>  [362,] -0.27002988  2.143743e-02  0.1571764124
#>  [363,] -0.28190404  1.961150e-02  0.1571608715
#>  [364,] -0.30329400  7.682623e-03  0.1774865554
#>  [365,] -0.28969044  1.795075e-03  0.1791964959
#>  [366,] -0.28837071  1.360917e-02  0.1718663902
#>  [367,] -0.29204441  1.458073e-02  0.1603522258
#>  [368,] -0.29387626  8.864123e-03  0.1657071402
#>  [369,] -0.27780456  2.298168e-02  0.2437521089
#>  [370,] -0.29892032  1.017564e-02  0.1708383520
#>  [371,] -0.25349635 -3.282695e-02  0.0978957223
#>  [372,] -0.26554897  2.950211e-03  0.1860965593
#>  [373,] -0.34528035 -6.129610e-02  0.0868150771
#>  [374,] -0.32869819  2.483824e-02  0.1383310330
#>  [375,] -0.32720863  1.259681e-02  0.1841180282
#>  [376,] -0.30086260  1.569552e-02  0.2010493110
#>  [377,] -0.29176878 -4.277720e-04  0.1860737689
#>  [378,] -0.31182230  5.038309e-03  0.1907754531
#>  [379,] -0.32277709 -4.600169e-03  0.1864999899
#>  [380,] -0.30903023 -1.101042e-04  0.1780765525
#>  [381,] -0.31393086  9.602332e-03  0.1815679868
#>  [382,] -0.30610723  2.180636e-02  0.1740223305
#>  [383,] -0.31843741  1.960793e-02  0.1761128142
#>  [384,] -0.33303831 -1.403582e-02  0.2095567861
#>  [385,] -0.29157667  1.066414e-02  0.1731852633
#>  [386,] -0.29325481  9.423554e-03  0.1607797805
#>  [387,] -0.29645298  2.463755e-02  0.1732916676
#>  [388,] -0.27984252  1.317257e-02  0.1652448304
#>  [389,] -0.28880847  4.160581e-02  0.1495549535
#>  [390,] -0.28225042  1.122464e-02  0.1522617996
#>  [391,] -0.30363401 -7.647216e-03  0.1449055697
#>  [392,] -0.31852293 -1.874336e-02  0.1758574620
#>  [393,] -0.28408444  3.829782e-02  0.1546403918
#>  [394,] -0.30964898  1.319667e-02  0.1728734226
#>  [395,] -0.28734921  1.578749e-02  0.1598106665
#>  [396,] -0.29482050  2.487458e-02  0.1650985165
#>  [397,] -0.31754052 -9.343129e-04  0.1913002402
#>  [398,] -0.29336959  2.446350e-02  0.1633948050
#>  [399,] -0.26263961  1.507296e-02  0.1674780871
#>  [400,] -0.32223546 -1.028026e-02  0.1947135477
#>  [401,] -0.30713946  1.218636e-02  0.1836031051
#>  [402,] -0.30588881  3.301976e-03  0.1765044977
#>  [403,] -0.31948096  1.021646e-02  0.1825347439
#>  [404,] -0.27312735  1.846173e-02  0.1654364054
#>  [405,] -0.29555877  2.464576e-02  0.1548995472
#>  [406,] -0.29757742 -3.314928e-02  0.1615733288
#>  [407,] -0.31466253 -6.816000e-03  0.1831847289
#>  [408,] -0.30181813  2.885231e-03  0.1753150685
#>  [409,] -0.29981649  2.945119e-02  0.1656669315
#>  [410,] -0.29369418 -4.754338e-02  0.2058750518
#>  [411,] -0.29108659 -2.798622e-02  0.2221808049
#>  [412,] -0.29629733  7.000216e-02  0.1393391225
#>  [413,] -0.29519899  1.365535e-02  0.1715395069
#>  [414,] -0.31558958 -4.130532e-03  0.1847649129
#>  [415,] -0.30038509  2.815880e-02  0.1599544436
#>  [416,] -0.30396334  3.073176e-02  0.1986852138
#>  [417,] -0.30180361  5.510648e-03  0.1805724296
#>  [418,] -0.28176005  1.720760e-02  0.1530079517
#>  [419,] -0.31276618 -5.547550e-03  0.1851922059
#>  [420,] -0.28277333  1.797897e-02  0.1722430879
#>  [421,] -0.28013107  2.746508e-02  0.1482614196
#>  [422,] -0.28876058  1.800385e-02  0.1651552370
#>  [423,] -0.32082189 -5.889374e-03  0.1784096909
#>  [424,] -0.27340378  2.154179e-02  0.1502324385
#>  [425,] -0.27725325  2.942184e-02  0.1536864779
#>  [426,] -0.25227173  3.193143e-03  0.1469070201
#>  [427,] -0.33576208 -5.932161e-02  0.1456226769
#>  [428,] -0.30108097  3.893745e-03  0.1783030572
#>  [429,] -0.31879147  2.448255e-03  0.1964293457
#>  [430,] -0.31533840 -1.251506e-03  0.1805135230
#>  [431,] -0.27314316  1.375592e-02  0.1308644801
#>  [432,] -0.27947716  2.426286e-02  0.1643198919
#>  [433,] -0.28765104 -2.338858e-03  0.1854928224
#>  [434,] -0.29290515  3.765889e-03  0.1719093161
#>  [435,] -0.29040521  1.412770e-02  0.1528372714
#>  [436,] -0.30101070  8.176279e-03  0.1812399401
#>  [437,] -0.28188333  1.272349e-02  0.1592132292
#>  [438,] -0.35440930  1.360337e-02  0.1297485569
#>  [439,] -0.28326014  2.619743e-02  0.1869125842
#>  [440,]  0.06194980  2.631296e-03  0.1700708061
#>  [441,] -0.32287417  4.512613e-03  0.1865529077
#>  [442,] -0.31917227 -1.447811e-03  0.1911875307
#>  [443,] -0.28463342  1.593208e-02  0.1444191892
#>  [444,] -0.29984469 -2.575129e-03  0.1680592458
#>  [445,] -0.31423673  7.255939e-03  0.1934085816
#>  [446,] -0.31280275  2.277502e-03  0.1849497001
#>  [447,] -0.31069270  2.380211e-03  0.1824448510
#>  [448,] -0.24378709  2.186310e-02  0.1324266906
#>  [449,] -0.27660016  2.677128e-02  0.1475134389
#>  [450,] -0.31192168 -6.336125e-03  0.1800982574
#>  [451,] -0.30390001  7.259022e-03  0.1747245364
#>  [452,] -0.27512406 -4.809972e-03  0.1859782975
#>  [453,] -0.28686506  4.868626e-02  0.2606473058
#>  [454,] -0.31556282  5.681407e-03  0.1792688525
#>  [455,] -0.29111070  1.295575e-02  0.1764123586
#>  [456,] -0.26725072  2.567258e-02  0.2110883810
#>  [457,] -0.28305964  5.624634e-02  0.1582834254
#>  [458,] -0.30642808  7.548899e-03  0.1837013571
#>  [459,] -0.27524545  2.175125e-02  0.1671917072
#>  [460,] -0.28722025  1.446241e-02  0.1563840359
#>  [461,] -0.28628156 -5.676898e-03  0.1849528171
#>  [462,] -0.29729988  1.146368e-02  0.1638080321
#>  [463,] -0.28235993  1.677581e-04  0.1451355780
#>  [464,] -0.26154983 -1.684261e-03  0.1420005261
#>  [465,] -0.27948152  2.196729e-02  0.1611265453
#>  [466,] -0.32884638  5.781377e-03  0.1877959424
#>  [467,] -0.30525846  1.749812e-02  0.1889924373
#>  [468,] -0.31393008 -1.842389e-03  0.1863986332
#>  [469,] -0.28523493  8.082584e-03  0.1608637370
#>  [470,] -0.27226013  2.466046e-02  0.1456226426
#>  [471,] -0.30134128  1.510155e-02  0.2132850768
#>  [472,] -0.26255993  1.978407e-02  0.1353702875
#>  [473,] -0.32439116 -1.108248e-03  0.1878885948
#>  [474,] -0.28272534  3.185995e-03  0.1658450498
#>  [475,] -0.28404019  4.400499e-03  0.1426521551
#>  [476,] -0.26218412 -8.884490e-02  0.1768543057
#>  [477,] -0.26797388  2.514287e-02  0.1451475453
#>  [478,] -0.31920289 -9.639763e-03  0.1896727839
#>  [479,] -0.30125683  2.025416e-02  0.1501853010
#>  [480,] -0.29307920  1.167442e-01  0.1648730151
#>  [481,] -0.30202234  1.568919e-02  0.1813985049
#>  [482,] -0.32410213 -7.900977e-03  0.1765975092
#>  [483,] -0.27800024  3.811801e-03  0.1593989377
#>  [484,] -0.30636232  1.689993e-02  0.1707356520
#>  [485,] -0.30610727 -1.195808e-02  0.1773667647
#>  [486,] -0.27179644 -5.140234e-02 -0.2290873725
#>  [487,] -0.29860076 -2.629807e-03  0.1686121491
#>  [488,] -0.24171467  2.238890e-02  0.1278516022
#>  [489,] -0.34558109 -5.956504e-03  0.1658018808
#>  [490,] -0.32405672 -3.597536e-03  0.1922221622
#>  [491,] -0.27943810  1.609525e-02  0.1580791877
#>  [492,] -0.29662090  3.696645e-02  0.1518454239
#>  [493,] -0.29395972 -8.736744e-03  0.1668456673
#>  [494,] -0.32844583 -9.125505e-03  0.2071298553
#>  [495,] -0.30152399 -2.936042e-02  0.1527287628
#>  [496,] -0.28222813 -1.766753e-02  0.0789015044
#>  [497,] -0.30221458  1.133784e-02  0.1576554192
#>  [498,] -0.27254518  1.388741e-02  0.1447918111
#>  [499,] -0.29655372  1.184446e-02  0.1688720886
#>  [500,] -0.32358872 -1.562447e-02  0.1885679776
#>  [501,] -0.21433963 -4.904756e-03  0.1753525391
#>  [502,] -0.28794570  1.808311e-02  0.1863045749
#>  [503,] -0.29034592  4.501968e-02  0.1841401729
#>  [504,] -0.30355293  8.559748e-03  0.1848797633
#>  [505,] -0.30666444  4.730190e-04  0.1686910069
#>  [506,] -0.28590680  1.786005e-02  0.1507149668
#>  [507,] -0.30005232  5.593850e-03  0.1655836697
#>  [508,] -0.32653474 -6.334719e-03  0.1997092677
#>  [509,] -0.30112215  8.070179e-03  0.1757829520
#>  [510,] -0.25645016  1.858762e-02  0.1232576327
#>  [511,] -0.26863290  2.332404e-02  0.1385673991
#>  [512,] -0.33339917 -7.315090e-03  0.1835103868
#>  [513,] -0.31711562 -1.747736e-03  0.1878163734
#>  [514,] -0.30695042  7.369327e-03  0.1810737608
#>  [515,] -0.29423473  1.421680e-02  0.1756063932
#>  [516,] -0.29452722  1.293117e-02  0.1632844095
#>  [517,] -0.23136345 -2.261458e-02  0.1641877236
#>  [518,] -0.26242199  5.072433e-02  0.1442111593
#>  [519,] -0.28001884  1.217858e-02  0.1425862662
#>  [520,] -0.28638742  7.563137e-03  0.1585206170
#>  [521,] -0.30534850 -6.106566e-04  0.1742515072
#>  [522,] -0.29982368  2.408518e-03  0.1533230493
#>  [523,] -0.28623779  5.960538e-03  0.1704551163
#>  [524,] -0.28488436 -1.837334e-03  0.1692380029
#>  [525,] -0.27170593  1.593156e-02  0.1537029305
#>  [526,] -0.26969205 -2.733631e-02  0.1374538376
#>  [527,] -0.28134431  1.745591e-04  0.1479292155
#>  [528,] -0.29195988  5.147067e-02  0.1397699954
#>  [529,] -0.06576525  7.561167e-03  0.1137300451
#>  [530,] -0.28379961  2.092170e-02  0.1614038867
#>  [531,] -0.28951504 -3.703433e-02  0.1945069926
#>  [532,] -0.26426483  1.731573e-02  0.1637354846
#>  [533,] -0.28495068  1.523919e-02  0.1984769164
#>  [534,] -0.31197435  6.476362e-03  0.1878393781
#>  [535,] -0.28465056  6.069535e-02  0.1488121014
#>  [536,] -0.31740819 -5.225664e-03  0.1867209992
#>  [537,] -0.30734422 -3.796165e-03  0.1802198751
#>  [538,] -0.26805977  4.893035e-02  0.0908065588
#>  [539,] -0.33049411 -7.244016e-03  0.1861880252
#>  [540,] -0.26964060  4.432818e-02  0.1519713566
#>  [541,] -0.32377529 -2.475879e-03  0.1928612476
#>  [542,] -0.28884792  2.402809e-02  0.1865013968
#>  [543,] -0.28245078 -4.127857e-03  0.0801889463
#>  [544,] -0.27534789  7.297599e-03  0.1402619559
#>  [545,] -0.31555602  3.823222e-03  0.1766380128
#>  [546,] -0.31723170  1.077967e-02  0.1798645428
#>  [547,] -0.30993806  5.501160e-03  0.1851019794
#>  [548,] -0.29582659  1.023367e-02  0.1767521123
#>  [549,] -0.30100998 -5.730329e-03  0.1805114638
#>  [550,] -0.31988691  1.398506e-02  0.1507476316
#>  [551,] -0.28007331 -5.775494e-03  0.1287790895
#>  [552,] -0.22401193  2.190823e-02 -0.0009398283
#>  [553,] -0.32296626 -3.670416e-04  0.1968291721
#>  [554,] -0.32450617  1.790765e-03  0.1935809380
#>  [555,] -0.28101642  9.893468e-03  0.1701431281
#>  [556,] -0.31638305  2.192723e-03  0.1911502210
#>  [557,] -0.25826849 -3.607975e-02  0.0894246749
#>  [558,] -0.29033012  1.056476e-02  0.1680382948
#>  [559,] -0.25379501 -7.936787e-02  0.0633315171
#>  [560,] -0.25004988  6.567418e-03  0.1005403182
#>  [561,] -0.50000000 -5.564698e-04  0.1686082530
#>  [562,] -0.29405787  1.002887e-02  0.1655824769
#>  [563,] -0.31516775 -1.155242e-02  0.1595803059
#>  [564,] -0.28376516  7.384596e-03  0.1590987645
#>  [565,] -0.29069754  5.536118e-03  0.1656097718
#>  [566,] -0.28117050  1.286222e-02  0.1612421028
#>  [567,] -0.31159126  1.369801e-03  0.1586526505
#>  [568,] -0.28584315  1.128606e-02  0.1602934348
#>  [569,] -0.30126041  4.878865e-03  0.1748755515
#>  [570,] -0.28753023 -7.172390e-03  0.1804326274
#>  [571,] -0.29818412 -1.133831e-02  0.1743105402
#>  [572,] -0.29123887 -3.954810e-03  0.1712031366
#>  [573,] -0.26429051  1.571508e-02  0.1495134623
#>  [574,] -0.28471267 -2.399861e-03  0.1926348780
#>  [575,] -0.28229219  1.984626e-02  0.1742356337
#>  [576,] -0.31531371  1.195538e-02  0.1837841934
#>  [577,] -0.32631638 -7.421451e-03  0.1964397055
#>  [578,] -0.29424582  1.530091e-02  0.1720585473
#>  [579,] -0.30248154  6.055101e-02  0.1435451053
#>  [580,] -0.32549278  6.778856e-02  0.1494782977
#>  [581,] -0.27378105  1.721780e-02  0.1521885363
#>  [582,] -0.34003657  1.829761e-02  0.1720832335
#>  [583,] -0.30092408  1.052356e-02  0.2164486257
#>  [584,] -0.30560856  2.279723e-02  0.1596819835
#>  [585,] -0.28927957  1.572664e-02  0.1289923625
#>  [586,] -0.28661822 -5.598156e-03  0.1950302247
#>  [587,] -0.27558476  1.905514e-02  0.1516173475
#>  [588,] -0.27438988  1.355020e-02  0.1554405375
#>  [589,] -0.28530519  3.039730e-02  0.1515352905
#>  [590,] -0.28371504  9.361202e-03  0.1755291537
#>  [591,] -0.29666893  2.286562e-03  0.1759874973
#>  [592,] -0.26646857  1.794910e-02  0.1595636511
#>  [593,] -0.28072534 -2.471748e-03  0.1487078584
#>  [594,] -0.30319931 -6.160597e-03  0.1561578648
#>  [595,] -0.28317162  1.336823e-02  0.1576530239
#>  [596,] -0.26105180 -4.304835e-01  0.2830255263
#>  [597,] -0.28933494  1.970421e-02  0.1543376649
#>  [598,] -0.26941886  2.060573e-02  0.1311345194
#>  [599,] -0.30664520  7.796529e-03  0.1747802042
#>  [600,] -0.28534236  3.472148e-02  0.1725702013
#>  [601,] -0.29482743  2.109078e-02  0.1759311283
#>  [602,] -0.28795565 -1.079022e-02  0.1808836760
#>  [603,] -0.30201954  6.179097e-03  0.1788893383
#>  [604,] -0.30123142  1.016246e-02  0.1679143514
#>  [605,] -0.31177814  2.376211e-03  0.1823058006
#>  [606,] -0.30113943  4.213422e-02  0.2954641733
#>  [607,] -0.28163692  6.043248e-02  0.1222953984
#>  [608,] -0.29432257  2.555496e-02  0.1863785857
#>  [609,] -0.29349836 -1.481946e-02  0.1358691877
#>  [610,] -0.28671169  1.830349e-02  0.1502081786
#>  [611,] -0.31927930 -1.273471e-02  0.1977584466
#>  [612,] -0.25635875  1.204060e-02  0.1672292081
#>  [613,] -0.31649237  5.483729e-02  0.1714951734
#>  [614,] -0.32147905 -5.662265e-03  0.1928625196
#>  [615,] -0.30129803  3.745012e-02  0.2430715786
#>  [616,] -0.28121988  3.353180e-03  0.1587875680
#>  [617,] -0.28596788  2.078083e-02  0.1593656224
#>  [618,] -0.27279979  6.865637e-03  0.1424194686
#>  [619,] -0.29450833  6.289542e-03  0.1769905152
#>  [620,] -0.28945446  1.471675e-02  0.1758051828
#>  [621,] -0.32831728 -3.277040e-03  0.2000981379
#>  [622,] -0.31484973  1.011049e-03  0.1856914993
#>  [623,] -0.28819244  2.039225e-02  0.1650555276
#>  [624,] -0.30101629  6.448554e-02  0.1737263311
#>  [625,] -0.27826897  2.863884e-03  0.1861214585
#>  [626,] -0.27637619  4.342866e-02  0.1484268105
#>  [627,] -0.31716607  4.582036e-03  0.1781566550
#>  [628,] -0.32957574  1.031438e-02  0.2183466415
#>  [629,] -0.30837891  1.787504e-02  0.1824295658
#>  [630,] -0.30296073  1.450274e-02  0.2127488363
#>  [631,] -0.29712499  2.635285e-02  0.1695490903
#>  [632,] -0.25918124  2.156621e-02  0.1504333403
#>  [633,] -0.27439213  1.209801e-02  0.1554581607
#>  [634,] -0.30949364 -6.463547e-03  0.1848157674
#>  [635,] -0.31288310  1.148328e-02  0.1662677169
#>  [636,] -0.30471459  1.569292e-03  0.1851730421
#>  [637,] -0.30420922 -2.427838e-03  0.1979011978
#>  [638,] -0.30402015  2.889843e-03  0.1783584943
#>  [639,] -0.31363480 -1.683994e-03  0.1846022856
#>  [640,] -0.30572881 -6.700908e-03  0.1927584366
#>  [641,] -0.28497359  2.324529e-02  0.1514782524
#>  [642,] -0.25019719  1.978714e-02  0.1679339765
#>  [643,] -0.30469224  4.660290e-04  0.1664372605
#>  [644,] -0.27825703  1.444356e-02  0.1441875959
#>  [645,] -0.27034689  3.226379e-02  0.2015225288
#>  [646,] -0.30623389  4.393263e-03  0.1735466937
#>  [647,] -0.27828390  7.122871e-03  0.1766157443
#>  [648,] -0.18599768  2.645924e-02  0.1687057209
#>  [649,] -0.30136794  2.514838e-02  0.1665245386
#>  [650,] -0.31398729  1.359952e-02  0.1449931565
#>  [651,] -0.26272356  1.155520e-02  0.1785045618
#>  [652,] -0.28585281  1.940354e-02  0.1464377432
#>  [653,] -0.31488877 -2.028611e-03  0.1723971288
#>  [654,] -0.32104167 -1.306637e-03  0.1874737837
#>  [655,] -0.30789578 -1.715356e-03  0.1780240482
#>  [656,] -0.28308799  6.046364e-02  0.1413146421
#>  [657,] -0.28537458  3.207036e-02  0.1534243027
#>  [658,] -0.30820913  1.679479e-03  0.1733207949
#>  [659,] -0.28453906  2.454058e-02  0.1566649289
#>  [660,] -0.33477464 -1.378704e-02  0.1904458731
#>  [661,] -0.31179455  5.877390e-03  0.1693514541
#>  [662,] -0.30485588 -1.218529e-02  0.0967699156
#>  [663,] -0.33389943 -1.142315e-02  0.2051197870
#>  [664,] -0.29737833  6.524634e-03  0.1766975185
#>  [665,] -0.30061765  5.399301e-03  0.1686042855
#>  [666,] -0.27302734 -1.950681e-02  0.1111325876
#>  [667,] -0.30764403  6.201528e-03  0.1837995084
#>  [668,] -0.27498323  1.545790e-02  0.1475695564
#>  [669,] -0.26552606  3.285299e-02  0.1269717943
#>  [670,] -0.30708632 -1.602066e-02  0.1692492809
#>  [671,] -0.28164581  1.282752e-02  0.1549553716
#>  [672,] -0.28752541  8.976841e-03  0.1587909193
#>  [673,] -0.30122138  4.287685e-04  0.1676570310
#>  [674,] -0.29814808  3.817460e-02  0.1642506150
#>  [675,] -0.27441555  1.340020e-02  0.1663420936
#>  [676,] -0.29533813  6.211963e-03  0.1971940836
#>  [677,] -0.28696970  4.369337e-02  0.1510385571
#>  [678,] -0.24651027  7.231225e-02  0.2196977408
#>  [679,] -0.31506902 -1.097461e-02  0.1456006062
#>  [680,] -0.28341715  1.415219e-02  0.1591048517
#>  [681,] -0.30404370  2.000960e-02  0.1732486991
#>  [682,] -0.26356416  2.621557e-02  0.1743397017
#>  [683,] -0.32286133 -3.127806e-03  0.1837776144
#>  [684,] -0.30041101  3.848174e-02  0.1646006685
#>  [685,] -0.28074518  7.102899e-03  0.1485867797
#>  [686,] -0.29336015  1.718391e-02  0.1660048360
#>  [687,] -0.27360512  1.242702e-02  0.1661657689
#>  [688,] -0.28356904 -6.862982e-03  0.0405856806
#>  [689,] -0.25829253  3.423772e-02  0.1915942742
#>  [690,] -0.28676090  1.029865e-02  0.1492829089
#>  [691,] -0.31752058 -6.710780e-04  0.1905460111
#>  [692,] -0.29786382  1.756466e-02  0.1660129760
#>  [693,] -0.30506372  4.098881e-03  0.1699869140
#>  [694,] -0.29261437  6.656119e-03  0.1456026433
#>  [695,] -0.26820492  1.017569e-02  0.1388237681
#>  [696,] -0.25840101  3.795161e-02  0.1328850858
#>  [697,] -0.25542973 -9.637075e-02 -0.3609551769
#>  [698,] -0.30218740  8.289449e-03  0.1644151967
#>  [699,] -0.29218179  3.362329e-03  0.1700982889
#>  [700,] -0.28461163  2.234286e-02  0.1352124370
#>  [701,] -0.30731700 -3.633565e-04  0.1853037878
#>  [702,] -0.30222429  6.219541e-03  0.1764380985
#>  [703,] -0.28320529  4.015923e-03  0.1603013731
#>  [704,] -0.30789226  4.049963e-02  0.1985318948
#>  [705,] -0.30804627  4.073741e-03  0.2115718054
#>  [706,] -0.28655206 -7.224478e-03  0.1527805753
#>  [707,] -0.28004157  1.205123e-02  0.1588683138
#>  [708,] -0.28093513  1.376084e-02  0.1562629572
#>  [709,] -0.30782104  2.859851e-03  0.1837000651
#>  [710,] -0.31998003 -4.191740e-03  0.2117178596
#>  [711,] -0.30013698  2.765352e-02  0.1813686578
#>  [712,] -0.31524654 -3.028942e-03  0.1888562113
#>  [713,] -0.29313272  2.450256e-02  0.1658815664
#>  [714,] -0.29893994  1.317598e-02  0.2274115664
#>  [715,] -0.30857613 -2.767329e-03  0.1919668120
#>  [716,] -0.32746296 -5.763505e-03  0.2005096810
#>  [717,] -0.29527169  4.806550e-02  0.1713819873
#>  [718,] -0.28714683  5.905092e-02  0.1230026684
#>  [719,] -0.28487515  9.714755e-03  0.1533138640
#>  [720,] -0.28132215 -1.634716e-02  0.1544309216
#>  [721,] -0.28021556  1.470302e-02  0.1200042039
#>  [722,] -0.28979244  5.680577e-03  0.0915175401
#>  [723,] -0.30875692  3.878830e-03  0.1775096695
#>  [724,] -0.29816933 -4.868877e-02  0.3609551769
#>  [725,] -0.32681595 -3.914290e-03  0.1930359350
#>  [726,] -0.33561643 -1.802650e-02  0.2075818024
#>  [727,] -0.29008502  1.024664e-02  0.1719238750
#>  [728,] -0.30703077  1.264272e-02  0.1908163721
#>  [729,] -0.29286031  2.805879e-03  0.1615197984
#>  [730,] -0.29854628  1.336486e-02  0.1701992926
#>  [731,] -0.32561218  1.942530e-03  0.1867991274
#>  [732,] -0.26652583  2.857157e-02  0.1306289481
#>  [733,] -0.27791404  3.615174e-02  0.1527987711
#>  [734,] -0.30531811  1.445333e-03  0.1875390929
#>  [735,] -0.28502415  4.032698e-02  0.1516387454
#>  [736,] -0.28275942  8.480076e-02  0.1811712334
#>  [737,] -0.28078662  1.160048e-02  0.1550187314
#>  [738,] -0.29368040  1.364820e-03  0.1640679990
#>  [739,] -0.29048442  3.229836e-02  0.1613201506
#>  [740,] -0.31595879  2.430917e-02  0.1748320398
#>  [741,] -0.28848840  1.409210e-02  0.1729374631
#>  [742,] -0.27557077  2.372451e-02  0.1472535883
#>  [743,] -0.27640239  2.362546e-02  0.1485000270
#>  [744,] -0.30232550  8.057200e-06  0.1697214588
#>  [745,] -0.29598327  7.130935e-03  0.1801611460
#>  [746,] -0.27663504  1.036469e-03  0.1509411332
#>  [747,] -0.29202449  3.362494e-02  0.1494515783
#>  [748,] -0.28846801  9.930355e-03  0.1653439740
#>  [749,] -0.32310522 -3.864396e-03  0.2021354315
#>  [750,] -0.30578353 -1.076166e-04  0.1702435378
#>  [751,] -0.28632549 -3.525717e-03  0.0257347931
#>  [752,] -0.31427087  1.937763e-03  0.1853814208
#>  [753,] -0.26638392  2.136569e-02  0.1393555766
#>  [754,] -0.29737564  9.218363e-03  0.1722710469
#>  [755,] -0.29310915  1.319319e-02  0.1804975382
#>  [756,] -0.27883262  3.263927e-02  0.1447654880
#>  [757,] -0.29147852  1.799693e-03  0.2014403280
#>  [758,] -0.28962145  1.334782e-02  0.1537970890
#>  [759,] -0.31773561 -1.039456e-02  0.1875173915
#>  [760,] -0.29691247  1.170297e-02  0.1650515189
#>  [761,] -0.28282071  1.095651e-02  0.1560205192
#>  [762,] -0.27234240  3.661116e-03  0.1396165941
#>  [763,] -0.38334139 -1.762040e-02  0.1489672476
#>  [764,] -0.30372735  3.240586e-02  0.1672831903
#>  [765,] -0.28409286  7.046356e-02  0.1780497667
#>  [766,] -0.28558528  1.604934e-02  0.1700476707
#>  [767,] -0.29232159 -1.520526e-02  0.1698731514
#>  [768,] -0.29383846  1.421973e-02  0.1416846929
#>  [769,] -0.29034531  8.995615e-03  0.1381236159
#>  [770,] -0.28100577  1.951106e-02  0.1504110457
#>  [771,] -0.33094606 -1.644544e-02  0.1995847618
#>  [772,] -0.28845344 -7.571805e-03  0.1686147780
#>  [773,] -0.27928134  2.260386e-02  0.1534175817
#>  [774,] -0.29745652 -7.739353e-03  0.1549123901
#>  [775,] -0.31037844 -1.416903e-02  0.1925630433
#>  [776,] -0.30980254  1.434317e-02  0.1713488227
#>  [777,] -0.29566505 -1.365571e-03  0.1699596913
#>  [778,] -0.31649028 -9.083687e-03  0.1924411813
#>  [779,] -0.30124006  2.157966e-02  0.1692647455
#>  [780,] -0.29056364  6.454116e-03  0.1787648520
#>  [781,] -0.28853483  9.479603e-02  0.1219932758
#>  [782,] -0.25805895  2.807107e-02  0.1417136620
#>  [783,] -0.30864267  4.454517e-02  0.1659249722
#>  [784,] -0.34826291 -7.364488e-03  0.1896685872
#>  [785,] -0.29811446  1.223449e-02  0.1623771484
#>  [786,] -0.27114556  4.936876e-03  0.1615831960
#>  [787,] -0.31858556 -2.734675e-03  0.1883831421
#>  [788,] -0.28283805  1.707542e-02  0.1566824729
#>  [789,] -0.29811509  5.332350e-03  0.1806111719
#>  [790,] -0.28547624  1.616021e-02  0.1646779826
#>  [791,] -0.32089218 -2.593469e-03  0.1924844366
#>  [792,] -0.30696344 -1.402653e-02  0.1754543428
#>  [793,] -0.30265647 -1.841310e-02  0.1858370850
#>  [794,] -0.30876580 -2.198381e-03  0.1843909002
#>  [795,] -0.34225354 -5.732165e-04  0.1920167731
#>  [796,] -0.30584873 -7.107926e-03  0.1913423374
#>  [797,] -0.28694847  7.025696e-02  0.1915181109
#>  [798,] -0.30249771 -1.245254e-03  0.1678560731
#>  [799,] -0.29158354 -1.049624e-02  0.1886377578
#>  [800,] -0.26931722  2.421760e-01  0.0849081569
#>  [801,] -0.32806616 -6.998651e-03  0.2024713039
#>  [802,] -0.31614330 -8.606145e-04  0.1875204818
#>  [803,] -0.33223677  1.349661e-04  0.2070422854
#>  [804,] -0.28299258  4.459953e-02  0.1413441150
#>  [805,] -0.23360152  7.889075e-02  0.1384209093
#>  [806,] -0.29818216 -5.160975e-03  0.1893119862
#>  [807,] -0.28308195  1.279861e-02  0.1609940783
#>  [808,] -0.29887096 -7.515555e-03  0.2096288781
#>  [809,] -0.30594621  3.132369e-04  0.1784425154
#>  [810,] -0.28028875  1.909115e-02  0.1617247242
#>  [811,] -0.28911346  2.214396e-02  0.1578322246
#>  [812,] -0.30614742 -1.348248e-02  0.1800406181
#>  [813,] -0.29355593  1.612185e-02  0.1651630586
#>  [814,] -0.30854080  9.303251e-03  0.1665184142
#>  [815,] -0.29008981  9.362554e-03  0.1793490545
#>  [816,] -0.28171881  2.392951e-02  0.1442430848
#>  [817,] -0.30920511  3.784858e-03  0.1833083341
#>  [818,] -0.29699038  3.230491e-02  0.1381149633
#>  [819,] -0.28746293  1.511210e-03  0.1713071641
#>  [820,] -0.28805515  2.153692e-03  0.2576300450
#>  [821,] -0.30839029  5.480960e-03  0.1832500055
#>  [822,] -0.28319365  3.275899e-02  0.1440281330
#>  [823,] -0.28984979  1.918953e-02  0.1622368600
#>  [824,] -0.32306990 -5.224691e-03  0.1899997819
#>  [825,] -0.28784451  7.282979e-03  0.1665698261
#>  [826,] -0.29369198  5.533085e-03  0.1615819936
#>  [827,] -0.30044364  9.763760e-03  0.1681550079
#>  [828,] -0.31360085  1.716997e-02  0.1859761902
#>  [829,] -0.31084788  6.420995e-04  0.1926997956
#>  [830,] -0.28840038  3.842095e-03  0.1536161355
#>  [831,] -0.31191031  7.556242e-03  0.1761798568
#>  [832,] -0.30541664  8.873020e-03  0.1409498507
#>  [833,] -0.32315348  2.478420e-03  0.1719179115
#>  [834,] -0.32035634  3.389765e-03  0.1964190096
#>  [835,] -0.29604764  2.708453e-02  0.1501440568
#>  [836,] -0.27052059  4.838577e-02  0.1284849398
#>  [837,] -0.27341226  2.090498e-02  0.1577045353
#>  [838,] -0.32160803  2.046561e-03  0.1903592674
#>  [839,] -0.30584646  2.233900e-02  0.1889221749
#>  [840,] -0.27746445  2.242062e-02  0.1569462682
#>  [841,] -0.22666442 -3.152934e-01 -0.1748213726
#>  [842,] -0.28039983  1.937338e-03  0.2033678665
#>  [843,] -0.31027218 -4.094178e-03  0.1747020407
#>  [844,] -0.30397777  5.401737e-02  0.1491070579
#>  [845,] -0.29862420  8.194059e-03  0.1775261965
#>  [846,] -0.29186892  6.269823e-03  0.1674379267
#>  [847,] -0.30444380  3.963177e-02  0.1559895048
#>  [848,] -0.32583688 -2.608082e-03  0.1755852791
#>  [849,] -0.30895887  1.116429e-02  0.1703842777
#>  [850,] -0.28285756  9.091945e-03  0.1683763311
#>  [851,] -0.29697558  1.481549e-02  0.1676660333
#>  [852,] -0.27968744  3.646686e-02  0.1686411466
#>  [853,] -0.27064415  1.498371e-02  0.1567193136
#>  [854,] -0.30714800  9.654706e-03  0.1766890650
#>  [855,] -0.30950726 -6.330452e-02  0.2207414049
#>  [856,] -0.32710553 -1.808272e-02  0.1718275692
#>  [857,] -0.29247196 -1.706503e-02 -0.0514276923
#>  [858,] -0.26571295  6.401603e-03  0.1562431191
#>  [859,] -0.28960622  1.511795e-03  0.1650136704
#>  [860,] -0.17382162  4.731500e-02  0.2090748013
#>  [861,] -0.29990278  1.093936e-02  0.1710762175
#>  [862,] -0.30694212  4.104978e-03  0.1800713801
#>  [863,] -0.31482304  8.697538e-03  0.1914399578
#>  [864,] -0.28538766  5.900585e-03  0.1512240714
#>  [865,] -0.28498624  1.010005e-02  0.1687748336
#>  [866,] -0.30362906  7.153651e-03  0.1792837070
#>  [867,] -0.29897669  6.734592e-03  0.1699068839
#>  [868,] -0.28788943  6.297362e-03  0.1700921222
#>  [869,] -0.29623246  1.137757e-02  0.1711896504
#>  [870,] -0.32145317 -4.661240e-02  0.1891799418
#>  [871,] -0.31886964 -6.575137e-03  0.1923679550
#>  [872,] -0.30272478  4.087111e-02  0.1625323612
#>  [873,] -0.30103734  5.623138e-03  0.1696033891
#>  [874,] -0.29176482  1.663247e-03  0.1651569333
#>  [875,] -0.31911760 -2.490932e-05  0.1853712760
#>  [876,] -0.28215801 -6.244024e-03  0.1669030859
#>  [877,] -0.26046946  1.457843e-01  0.1103759370
#>  [878,] -0.30722918  3.166547e-03  0.1707898699
#>  [879,] -0.31269317 -1.811654e-03  0.1922574182
#>  [880,] -0.32273840 -1.071401e-04  0.1959694044
#>  [881,] -0.29009410  7.716988e-03  0.1634518533
#>  [882,] -0.25571610  3.804388e-02  0.1310631504
#>  [883,] -0.26690742  2.343759e-02  0.1442707106
#>  [884,] -0.30853265  2.182663e-03  0.1881832362
#>  [885,] -0.30275059  2.836160e-03  0.1511201753
#>  [886,] -0.29958831  3.711082e-04  0.1753657924
#>  [887,] -0.25837255  3.272514e-02  0.2425415631
#>  [888,] -0.27669920 -1.824137e-02  0.2016764862
#>  [889,] -0.30585158  1.865290e-02  0.1698764161
#>  [890,] -0.32909515 -2.872024e-03  0.2046004143
#>  [891,] -0.32140319  3.543302e-03  0.1896513971
#>  [892,] -0.29063135  2.263550e-02  0.1560939212
#>  [893,] -0.30453957  6.038192e-03  0.1724707006
#>  [894,] -0.33374841 -1.675653e-02  0.2084675983
#>  [895,] -0.31350218 -2.063570e-03  0.1828336992
#>  [896,] -0.31947475  1.013677e-02  0.2008183679
#>  [897,] -0.30295429  8.966670e-03  0.1791751512
#>  [898,] -0.31926720  1.785278e-03  0.1895648117
#>  [899,] -0.28906216  3.577480e-03  0.1711959433
#>  [900,] -0.31980736 -3.081471e-03  0.1893090356
#>  [901,] -0.31595265  1.051234e-02  0.1173559577
#>  [902,] -0.31285182 -1.245857e-03  0.1907007169
#>  [903,] -0.29992204 -3.707743e-03  0.1789407063
#>  [904,] -0.29993374  8.815252e-03  0.1921316172
#>  [905,] -0.30250466  3.380963e-03  0.1697890541
#>  [906,] -0.28271500  5.058493e-02  0.1381662856
#>  [907,] -0.32502717  2.075643e-03  0.1951323928
#>  [908,] -0.26089383  1.396749e-02  0.1260382136
#>  [909,] -0.28799614 -2.843918e-03  0.1625056115
#>  [910,] -0.32132184  9.504479e-03  0.1621814420
#>  [911,] -0.32051755  1.388184e-03  0.1903701084
#>  [912,] -0.35527669 -2.096881e-02  0.1635436671
#>  [913,] -0.26726603  1.925863e-02  0.1576890219
#>  [914,] -0.34250649 -1.502574e-02  0.2109965358
#>  [915,] -0.30054958  2.925353e-02  0.1622885305
#>  [916,] -0.31991963 -4.620920e-03  0.1879556785
#>  [917,] -0.30334206  1.289212e-02  0.1753426738
#>  [918,] -0.27965385 -5.421210e-02  0.1829210558
#>  [919,] -0.31855513 -1.570027e-02  0.2174277709
#>  [920,] -0.30126560  1.066774e-02  0.1642994803
#>  [921,] -0.29037182 -8.998140e-03  0.1648721692
#>  [922,] -0.29268350  5.638451e-02  0.1715653844
#>  [923,] -0.29839712  1.000667e-02  0.1630923706
#>  [924,] -0.31083767 -7.972228e-04  0.1803729290
#>  [925,] -0.27086816  1.380582e-02  0.1551771645
#>  [926,] -0.29520192  9.866960e-03  0.1606256925
#>  [927,] -0.30472676  2.003178e-02  0.2507356529
#>  [928,] -0.31239168 -3.345899e-02  0.2170064545
#>  [929,] -0.31503890  1.587993e-03  0.1828485806
#>  [930,] -0.27231513  2.024094e-02  0.1545158713
#>  [931,] -0.30066332 -1.000004e-03  0.1792982218
#>  [932,] -0.27044266  9.008046e-04  0.1261170320
#>  [933,] -0.30485072  4.432654e-03  0.1764301580
#>  [934,] -0.33554348 -9.790759e-03  0.1984536346
#>  [935,] -0.29155104  1.202014e-02  0.2044692875
#>  [936,] -0.29624373  2.109625e-02  0.1770960448
#>  [937,] -0.30098197  1.002129e-03  0.1877366471
#>  [938,] -0.30652232 -1.800039e-03  0.1766983379
#>  [939,] -0.25619510  1.872779e-02  0.1696969905
#>  [940,] -0.28055860  1.603939e-02  0.1579292737
#>  [941,] -0.27066978  6.101814e-02  0.1306264150
#>  [942,] -0.27232298  3.799496e-02  0.1346600843
#>  [943,] -0.30960403  1.689700e-02  0.1855905647
#>  [944,] -0.28516319  1.547274e-02  0.1428829351
#>  [945,] -0.29205324  5.849156e-04  0.1672794334
#>  [946,] -0.28727229  5.576928e-02  0.1899493635
#>  [947,] -0.28796201  1.151880e-02  0.1456576780
#>  [948,] -0.31141990  1.590618e-03  0.1796743763
#>  [949,] -0.30053648  9.251528e-03  0.1761699973
#>  [950,] -0.30697040  6.759708e-04  0.1827572379
#>  [951,] -0.29154341  1.064402e-02  0.1521310314
#>  [952,] -0.31133707  1.543470e-02  0.1581727353
#>  [953,] -0.31477823  1.872201e-02  0.1537165598
#>  [954,] -0.29148054  1.313433e-02  0.1623082278
#>  [955,] -0.29073607  1.807220e-02  0.1721017012
#>  [956,] -0.28319041  1.156653e-02  0.1589442888
#>  [957,] -0.28918156  1.610619e-02  0.1518302238
#>  [958,] -0.30374182 -9.416359e-03  0.1913461818
#>  [959,] -0.26533308  9.816909e-03  0.1700581586
#>  [960,] -0.26303497  1.766743e-01  0.0207936661
#>  [961,] -0.28581245  1.466315e-02  0.1634376701
#>  [962,] -0.29356560  1.337161e-02  0.1643802067
#>  [963,] -0.28476304  1.762017e-02  0.1582153962
#>  [964,] -0.29613752  7.232745e-03  0.1683190416
#>  [965,]  0.03261894  7.286922e-03  0.1766700942
#>  [966,] -0.28808112  6.814158e-03  0.1637513848
#>  [967,] -0.31857153 -2.459534e-03  0.1914384400
#>  [968,] -0.30087956 -1.410621e-02  0.1479245598
#>  [969,] -0.31586415 -2.038864e-03  0.1839161564
#>  [970,] -0.32762234 -6.641926e-03  0.2080069638
#>  [971,] -0.29374315  8.607387e-03  0.1613713448
#>  [972,] -0.31421451 -8.384583e-03  0.1791159882
#>  [973,] -0.28196000  5.014903e-03  0.1508289919
#>  [974,] -0.26458835 -5.083558e-03  0.1555653673
#>  [975,] -0.31057177  5.307720e-03  0.1861250713
#>  [976,] -0.29467488  1.076639e-02  0.1679218101
#>  [977,] -0.26646946  9.015638e-03  0.1491521501
#>  [978,] -0.31010707  5.468515e-03  0.1651272909
#>  [979,] -0.26892585  2.132345e-02  0.1463612395
#>  [980,] -0.29645871  7.558994e-03  0.1688701190
#>  [981,] -0.31150638  1.127529e-02  0.1780769117
#>  [982,] -0.28953583  6.158610e-03  0.1591053715
#>  [983,] -0.33918438 -1.163712e-02  0.2064141411
#>  [984,] -0.31822747 -1.277103e-02  0.1811581618
#>  [985,] -0.27442983  2.421104e-02  0.1455808729
#>  [986,] -0.31474613  1.930331e-02  0.2649550012
#>  [987,] -0.31224425 -7.480724e-03  0.1996276227
#>  [988,] -0.27565458  1.027802e-02  0.1500793459
#>  [989,] -0.29916635  7.022340e-04  0.1717465322
#>  [990,] -0.28312484  9.821014e-03  0.1664119480
#>  [991,] -0.28618452  1.007393e-01  0.1209776697
#>  [992,] -0.30321922 -7.373883e-04  0.1176231251
#>  [993,] -0.25515937  3.548502e-02  0.1335616939
#>  [994,] -0.28194988  3.057189e-03  0.1472143161
#>  [995,] -0.28792052  2.125747e-02  0.1603406828
#>  [996,] -0.31266396 -7.020875e-03  0.1849578174
#>  [997,] -0.29241497 -8.800712e-02  0.2217498448
#>  [998,] -0.30568623  1.106616e-02  0.1722353202
#>  [999,] -0.31121958 -8.534510e-03  0.1869930418
#> [1000,] -0.32607281 -1.171131e-02  0.1900999309
#> attr(,"addend")
#> [1] -0.09528698  0.06480621  0.02327741
#> attr(,"multiplier")
#> [1] 1.350224