Landmark MDS

A fast dimensionality reduction method scaleable to large numbers of samples. Landmark Multi-Dimensional Scaling (LMDS) is an extension of classical 'Torgerson MDS', but rather than calculating a complete distance matrix between all pairs of samples, only the distances between a set of landmarks and the samples are calculated.

A fast dimensionality reduction method scaleable to large numbers of samples. Landmark Multi-Dimensional Scaling (LMDS) is an extension of classical Torgerson MDS´, but rather than calculating a complete distance matrix between all pairs of samples, only the distances between a set of landmarks and the samples are calculated.

Usage

lmds(
  x,
  ndim = 3,
  distance_method = c("euclidean", "pearson", "spearman", "cosine", "chisquared",
    "hamming", "kullback", "manhattan", "maximum", "canberra", "minkowski"),
  landmark_method = c("sample"),
  num_landmarks = 500
)

Arguments

x

A matrix, optionally sparse.

ndim

The number of dimensions

distance_method

The distance metric to use. Options are "euclidean" (default), "pearson", "spearman", "cosine", "manhattan".

landmark_method

The landmark selection method to use. Options are "sample" (default).

num_landmarks

The number of landmarks to use,

Value

The dimensionality reduction in the form of a nrow(x) by ndim matrix.

Examples

library(Matrix)
x <- Matrix::rsparsematrix(1000, 1000, .01)
lmds(x, ndim = 3)
#>              comp_1        comp_2        comp_3
#>    [1,]  0.21951873  8.805337e-02  1.067764e-02
#>    [2,]  0.20984638  7.490549e-02 -1.335701e-02
#>    [3,]  0.19994076  8.584810e-02 -2.768658e-02
#>    [4,]  0.20544634  7.370154e-02  9.095236e-03
#>    [5,]  0.17369673  6.910359e-02 -1.483816e-02
#>    [6,]  0.23289235  9.359322e-02 -1.512714e-02
#>    [7,]  0.18931591  6.592413e-02 -2.247413e-02
#>    [8,]  0.21920080  1.189694e-01 -2.273118e-01
#>    [9,]  0.21121751  9.178697e-02 -1.854893e-03
#>   [10,]  0.21414526  1.198990e-01 -9.391860e-03
#>   [11,]  0.19760598  6.691914e-02 -1.474273e-02
#>   [12,]  0.20985915  8.571092e-02 -1.572999e-02
#>   [13,]  0.20205568  7.112192e-02 -2.807911e-02
#>   [14,]  0.23946774  1.014085e-01 -1.522689e-02
#>   [15,]  0.19657909  6.822980e-02 -2.070016e-02
#>   [16,]  0.22024630  8.783948e-02 -1.365769e-02
#>   [17,]  0.18881325  8.887379e-02 -1.265011e-02
#>   [18,]  0.20097673  6.709120e-02 -2.950271e-02
#>   [19,]  0.23093053  1.019371e-01 -5.347534e-02
#>   [20,]  0.22734189  9.408058e-02 -1.082439e-02
#>   [21,]  0.21006142  8.660990e-02 -9.538522e-03
#>   [22,]  0.22525951  8.891857e-02 -1.410479e-02
#>   [23,]  0.20172557  9.988766e-02 -1.999473e-02
#>   [24,]  0.20036178  9.170372e-02 -6.459707e-03
#>   [25,]  0.23050735  9.043285e-02 -2.254384e-03
#>   [26,]  0.20852164  7.295006e-02 -7.476725e-02
#>   [27,]  0.16713600  6.531959e-02  1.991373e-02
#>   [28,]  0.23493424  9.434673e-02 -4.369934e-03
#>   [29,]  0.20593083  8.638279e-02 -3.366879e-03
#>   [30,]  0.20171205  8.244007e-02 -8.689515e-03
#>   [31,]  0.23654292 -1.574474e-03 -3.122894e-02
#>   [32,]  0.20333471  7.856671e-02 -1.653591e-02
#>   [33,]  0.20441804  8.099535e-02 -1.233339e-02
#>   [34,]  0.22164164  9.408078e-02 -1.650658e-02
#>   [35,]  0.20654008  5.209343e-02 -1.297717e-02
#>   [36,]  0.20895301 -2.529519e-02 -4.723518e-02
#>   [37,]  0.19505140  6.679762e-02 -2.497564e-02
#>   [38,]  0.23124900  9.391206e-02 -1.078009e-02
#>   [39,]  0.24118701  1.045973e-01 -6.802295e-03
#>   [40,]  0.22448434  9.546392e-02 -1.179313e-02
#>   [41,]  0.19652058  8.062461e-02 -1.652941e-02
#>   [42,]  0.18484415  5.271001e-02 -2.133037e-02
#>   [43,]  0.20707009  8.058581e-02 -1.564456e-02
#>   [44,]  0.20050946  8.723741e-02  3.367097e-03
#>   [45,]  0.23687653  9.133367e-02 -2.008574e-02
#>   [46,]  0.23703800  9.272446e-02  9.358524e-03
#>   [47,]  0.18557106  6.510985e-02 -2.192566e-02
#>   [48,]  0.21276587  7.907335e-02 -4.376980e-03
#>   [49,]  0.21532827  8.518840e-02 -9.794880e-03
#>   [50,]  0.19385364  8.806158e-02 -3.339996e-02
#>   [51,]  0.21910956  8.730616e-02 -1.450176e-02
#>   [52,]  0.23721155  1.000670e-01 -1.289726e-02
#>   [53,]  0.18308483  1.080207e-01 -3.475607e-02
#>   [54,]  0.20289973  7.481885e-02 -1.954113e-02
#>   [55,]  0.19224018  3.714069e-02 -3.241056e-02
#>   [56,]  0.19746471  6.754374e-02  1.246462e-02
#>   [57,]  0.23152848  7.934068e-02 -3.914533e-02
#>   [58,]  0.20527614  8.557937e-02 -1.040093e-02
#>   [59,]  0.21237872  5.560938e-02 -1.313978e-02
#>   [60,]  0.23830866  1.179319e-01  1.547283e-02
#>   [61,]  0.20693081  7.210217e-02 -1.568989e-02
#>   [62,]  0.23026231  8.985435e-02 -1.578645e-02
#>   [63,]  0.21133177  7.459083e-02 -1.489009e-02
#>   [64,]  0.23188419  9.629275e-02 -9.295043e-03
#>   [65,]  0.21906223  9.123279e-02 -1.106113e-02
#>   [66,]  0.21123825  8.743964e-02 -1.603341e-02
#>   [67,]  0.20470488  5.153192e-02 -1.987792e-02
#>   [68,]  0.21784260  6.485827e-02 -3.859140e-02
#>   [69,]  0.17429129  5.840068e-02 -1.419881e-02
#>   [70,]  0.23285582  9.814101e-02 -9.197899e-03
#>   [71,]  0.17447351  4.613150e-02 -3.489956e-02
#>   [72,]  0.21669209  8.964646e-02 -6.994910e-03
#>   [73,]  0.18080091  3.982466e-02 -3.261386e-02
#>   [74,]  0.15802421  1.040475e-01  7.367645e-03
#>   [75,]  0.20693130  7.814312e-02  1.115303e-02
#>   [76,]  0.22231454  8.791239e-02 -4.298415e-03
#>   [77,]  0.21536867  7.594376e-02 -2.315715e-02
#>   [78,]  0.21758635  7.264541e-02 -3.879402e-03
#>   [79,]  0.22695582  8.389959e-02 -1.181485e-02
#>   [80,]  0.20706712  8.421699e-02 -1.481799e-02
#>   [81,]  0.15441473 -7.072123e-02  3.325087e-01
#>   [82,]  0.18252550  3.491568e-02 -7.119101e-02
#>   [83,]  0.23134273  9.094921e-02 -1.579456e-02
#>   [84,]  0.18418778  1.260495e-01 -5.057275e-02
#>   [85,]  0.21218913  8.602921e-02 -1.326919e-02
#>   [86,]  0.20079618  6.666585e-02 -4.865237e-02
#>   [87,]  0.20477803  8.136811e-02  1.027149e-02
#>   [88,]  0.19917081  8.432414e-02 -1.278819e-02
#>   [89,]  0.18807293  6.740822e-02 -1.355925e-02
#>   [90,]  0.21756195  7.650850e-02 -2.578305e-02
#>   [91,]  0.21752218  9.097001e-02 -1.393675e-02
#>   [92,]  0.20391272  7.163724e-02 -1.041134e-02
#>   [93,]  0.22037915  9.160647e-02 -1.100347e-02
#>   [94,]  0.20434460  9.354419e-02 -8.662470e-03
#>   [95,]  0.21780548  8.606195e-02 -1.656825e-02
#>   [96,]  0.20689560  6.596868e-02 -2.862627e-02
#>   [97,]  0.21706785  8.451173e-02 -3.272392e-03
#>   [98,]  0.23091500  8.622422e-02 -1.702059e-02
#>   [99,]  0.23994415  1.023202e-01 -1.112655e-02
#>  [100,]  0.17051688  5.432508e-02  1.611380e-02
#>  [101,]  0.20728252  7.804892e-02 -2.344778e-02
#>  [102,]  0.22946167  8.969020e-02 -1.966163e-02
#>  [103,]  0.24047356  9.419158e-02 -5.371843e-04
#>  [104,]  0.20238775  6.810331e-02 -1.518082e-02
#>  [105,]  0.21901370  5.470981e-02 -3.702036e-03
#>  [106,]  0.21992733  7.312095e-02 -1.854273e-02
#>  [107,]  0.24008124  9.969047e-02 -9.487021e-03
#>  [108,]  0.19230444  1.049393e-01 -1.332081e-02
#>  [109,]  0.20504042  9.251076e-02 -3.614493e-04
#>  [110,]  0.15578202 -1.151491e-02 -1.726853e-01
#>  [111,]  0.18696611  6.894572e-02 -1.153453e-02
#>  [112,]  0.20763507  7.213564e-02 -1.842759e-02
#>  [113,]  0.20053977  7.480891e-02 -6.636950e-03
#>  [114,]  0.21815664  8.942598e-02 -1.344456e-02
#>  [115,]  0.21454442  1.067965e-01 -1.761085e-02
#>  [116,]  0.22875744  7.989571e-02  1.831256e-02
#>  [117,]  0.20922620  7.413359e-02 -6.614797e-03
#>  [118,]  0.21780380  1.094150e-01 -8.783398e-02
#>  [119,]  0.20172920  5.586503e-02 -2.011568e-02
#>  [120,]  0.20230926  7.660296e-02 -5.536912e-03
#>  [121,]  0.22360599  6.601371e-02 -2.067164e-02
#>  [122,]  0.20656218  9.934294e-02  7.708071e-04
#>  [123,]  0.18256081  3.119388e-02  2.098514e-02
#>  [124,]  0.21834172  8.280233e-02 -5.354042e-03
#>  [125,]  0.23345670  9.240867e-02 -1.009653e-02
#>  [126,]  0.21064398  8.116350e-02 -1.082294e-02
#>  [127,]  0.20524859  7.600621e-02 -1.612403e-02
#>  [128,]  0.19690258  7.466459e-02 -2.395276e-02
#>  [129,]  0.17948235  5.527631e-02 -1.466420e-02
#>  [130,]  0.22656686  8.292081e-02 -7.845085e-04
#>  [131,]  0.20806565  7.684028e-02 -1.384951e-02
#>  [132,]  0.20166037  7.795325e-02 -1.542781e-02
#>  [133,]  0.22196192  8.538435e-02 -2.493180e-02
#>  [134,]  0.24209624  2.822787e-02 -3.624617e-02
#>  [135,]  0.22415844  1.018969e-01 -7.798129e-03
#>  [136,]  0.22537305  8.765729e-02 -1.103979e-02
#>  [137,]  0.21251925  7.845774e-02 -1.673869e-02
#>  [138,]  0.24350091  9.724215e-02 -3.296394e-02
#>  [139,]  0.03187041 -3.531364e-02 -4.756998e-02
#>  [140,]  0.20700524  7.997171e-02 -4.314291e-02
#>  [141,]  0.21836202  8.258021e-02 -3.124769e-03
#>  [142,]  0.23906239  5.941715e-02 -3.425974e-02
#>  [143,]  0.23028186  9.140529e-02 -7.919683e-03
#>  [144,]  0.21012079  8.714734e-02 -1.772167e-02
#>  [145,]  0.22899899  9.325704e-02 -1.116672e-02
#>  [146,]  0.19707935  7.112334e-02  2.289486e-04
#>  [147,]  0.22724793  9.254743e-02 -1.432724e-02
#>  [148,]  0.15668328  1.409418e-01 -9.191549e-02
#>  [149,]  0.21552452  8.984650e-02  9.305628e-03
#>  [150,]  0.17296483  1.146358e-01 -5.379832e-04
#>  [151,]  0.20763086  5.096792e-02  8.087162e-02
#>  [152,]  0.26124122  1.145944e-01  3.184283e-03
#>  [153,]  0.20341706  7.251090e-02 -7.342982e-03
#>  [154,]  0.19528307  8.004398e-02 -7.331864e-03
#>  [155,]  0.14653899  3.338549e-02 -1.276141e-01
#>  [156,]  0.19407839  7.088815e-02 -1.454526e-02
#>  [157,]  0.21259021  7.902729e-02 -6.014904e-04
#>  [158,]  0.23584237  8.838568e-02  1.321865e-02
#>  [159,]  0.20759112  6.854625e-02  3.836578e-02
#>  [160,]  0.21089648  9.188989e-02 -1.067868e-02
#>  [161,]  0.18628859  6.578859e-02  6.088159e-02
#>  [162,]  0.21071486  4.585180e-02 -5.079368e-02
#>  [163,]  0.21917571  8.264772e-02 -1.360896e-02
#>  [164,]  0.23281300  9.598588e-02 -1.132688e-02
#>  [165,]  0.17042518  4.523726e-02  1.289744e-02
#>  [166,]  0.18920265  7.803749e-02 -1.933176e-02
#>  [167,]  0.19920860  7.415142e-02 -1.538406e-02
#>  [168,]  0.24651744  7.371251e-02 -2.992762e-02
#>  [169,]  0.18371530  7.220712e-02  7.469107e-03
#>  [170,]  0.21502277  8.099905e-02 -2.655432e-02
#>  [171,]  0.20867720  7.784302e-02 -1.581756e-02
#>  [172,]  0.23948417  8.890837e-02 -1.471605e-02
#>  [173,]  0.22084305  7.660342e-02  2.714078e-03
#>  [174,]  0.18016930  2.084860e-01  1.538547e-01
#>  [175,]  0.20470544  7.103052e-02 -1.442189e-02
#>  [176,]  0.18426588  1.298535e-02  1.062394e-01
#>  [177,]  0.17633159  3.886503e-02 -3.491906e-02
#>  [178,]  0.20183726  8.339888e-02 -7.268544e-03
#>  [179,]  0.21787132  8.237229e-02 -1.195689e-02
#>  [180,]  0.19798682  9.007175e-02 -4.453707e-03
#>  [181,]  0.21370198  8.122342e-02 -2.686491e-02
#>  [182,]  0.20943271  7.686174e-02 -1.593018e-02
#>  [183,]  0.19146663  7.590943e-02 -6.910705e-03
#>  [184,]  0.18293964  1.107170e-01 -1.676938e-03
#>  [185,]  0.18503694  4.974030e-02 -8.499289e-03
#>  [186,]  0.21313190  8.150403e-02 -1.377183e-02
#>  [187,]  0.22219865  9.325675e-02 -4.426162e-02
#>  [188,]  0.22348150  8.537979e-02  1.645055e-03
#>  [189,]  0.16813230  5.287223e-02  7.819709e-02
#>  [190,]  0.19377359  1.562190e-03  1.167244e-02
#>  [191,]  0.19965459  1.127752e-01 -1.667947e-02
#>  [192,]  0.22460141  8.486694e-02 -2.216626e-04
#>  [193,]  0.20917195  7.486681e-02  6.837214e-03
#>  [194,]  0.19228436  6.824747e-02 -7.035965e-03
#>  [195,]  0.19477355  6.788076e-02 -1.922136e-02
#>  [196,]  0.19645191  1.216253e-01 -1.801286e-02
#>  [197,]  0.23299750  3.560938e-02  3.956879e-02
#>  [198,]  0.24948132  1.079268e-01 -1.061831e-02
#>  [199,]  0.18930112  7.824865e-02 -1.462729e-02
#>  [200,]  0.19922802  7.951464e-02 -1.314624e-02
#>  [201,]  0.21172583  9.450260e-02  2.461689e-02
#>  [202,]  0.20573831  7.509832e-02 -1.117778e-02
#>  [203,]  0.20521873  7.440812e-02 -9.977258e-03
#>  [204,]  0.26434620  1.048422e-01 -1.138837e-02
#>  [205,]  0.22688660  8.934975e-02 -9.307126e-03
#>  [206,]  0.18872590  8.074537e-02 -1.222437e-02
#>  [207,]  0.22046460  8.516962e-02 -1.021361e-02
#>  [208,]  0.20989819  4.035900e-02 -8.148961e-03
#>  [209,]  0.24855072  7.731961e-02 -1.871170e-02
#>  [210,]  0.20439783  6.955054e-02 -1.703774e-02
#>  [211,]  0.21005203  6.994951e-02 -8.442015e-03
#>  [212,]  0.24498326  1.072960e-01 -1.598885e-02
#>  [213,]  0.21078030  8.759299e-02 -2.113619e-03
#>  [214,]  0.23304317  8.163276e-02 -7.935772e-03
#>  [215,]  0.19329995  6.548543e-02 -1.036563e-02
#>  [216,]  0.20421789  7.605621e-02 -4.731668e-03
#>  [217,]  0.21145064  7.853650e-02 -1.446289e-02
#>  [218,]  0.20648667  7.452501e-02 -1.694322e-02
#>  [219,]  0.21805788  8.221795e-02 -1.904087e-03
#>  [220,]  0.21844361  8.766248e-02 -1.164598e-02
#>  [221,]  0.18757488  8.676946e-02 -5.154353e-02
#>  [222,]  0.19966384  9.473937e-02 -1.759861e-01
#>  [223,]  0.19392697  6.169197e-02 -1.831984e-02
#>  [224,]  0.21017575  7.960041e-02 -3.090059e-02
#>  [225,]  0.19542478  7.338170e-02 -1.541286e-02
#>  [226,]  0.20465932  6.496216e-02 -1.797483e-02
#>  [227,]  0.27122087  7.882363e-02 -2.499219e-02
#>  [228,]  0.19796215  6.778892e-02 -1.199166e-02
#>  [229,]  0.20171283  7.245072e-02 -9.164901e-03
#>  [230,]  0.18911365  6.847769e-02 -1.570109e-02
#>  [231,]  0.17016150 -1.543667e-02 -1.386983e-01
#>  [232,]  0.23415475  9.969282e-02 -1.065971e-02
#>  [233,]  0.21252863  8.377423e-02 -2.424981e-02
#>  [234,]  0.22147845  9.795792e-02 -5.136096e-03
#>  [235,]  0.23928590  1.006916e-01 -1.101017e-02
#>  [236,]  0.18135374  6.214812e-02 -1.747139e-02
#>  [237,]  0.18031920  1.171142e-01  1.104082e-04
#>  [238,]  0.20754736  7.417685e-02 -1.181257e-02
#>  [239,]  0.18574921  2.245682e-01  5.909420e-02
#>  [240,]  0.21637141  9.049337e-02 -1.199986e-03
#>  [241,]  0.21877885  9.397260e-02  1.684685e-02
#>  [242,]  0.23411634  7.684539e-02 -8.626123e-03
#>  [243,]  0.24026660  1.022329e-01 -1.526168e-02
#>  [244,]  0.22926465  8.440001e-02 -6.500135e-03
#>  [245,]  0.21911657  8.604406e-02 -9.662920e-03
#>  [246,]  0.22490082  8.129768e-02 -4.365263e-02
#>  [247,]  0.20978265  7.702286e-02 -2.843276e-03
#>  [248,]  0.20986262  7.339157e-02  7.555639e-04
#>  [249,]  0.21252525  8.767554e-02 -1.128923e-02
#>  [250,]  0.22428166  8.282588e-02 -1.187082e-02
#>  [251,]  0.20778163  8.031489e-02  6.046480e-03
#>  [252,]  0.19456186  6.948494e-02 -1.827906e-02
#>  [253,]  0.20459601  8.506594e-02  7.540440e-03
#>  [254,]  0.18603849  5.229318e-02 -6.234977e-03
#>  [255,]  0.18265322  6.659240e-02 -1.131520e-02
#>  [256,]  0.20566828  8.128181e-02 -3.508956e-03
#>  [257,]  0.21695164  8.110225e-02 -1.251510e-02
#>  [258,]  0.21269720  8.512004e-02 -3.369621e-02
#>  [259,]  0.21448566  8.783128e-02 -2.533724e-02
#>  [260,]  0.20354823  6.602119e-02 -7.793767e-02
#>  [261,]  0.22192609  6.841150e-02 -3.776275e-02
#>  [262,]  0.19490244  8.314672e-02 -3.557098e-02
#>  [263,]  0.22870932  5.495244e-02 -1.234835e-02
#>  [264,]  0.23347479  1.016888e-01 -6.828381e-03
#>  [265,]  0.22872415  9.590773e-02 -9.739931e-03
#>  [266,]  0.18460904  5.498256e-02 -3.053889e-02
#>  [267,]  0.22184320  5.556759e-02 -9.867026e-04
#>  [268,]  0.23797871  9.901452e-02 -3.185168e-02
#>  [269,]  0.19150823  1.152670e-01  1.545155e-02
#>  [270,]  0.19260560  6.848293e-02 -1.637381e-02
#>  [271,]  0.21553433  7.530465e-02 -1.696519e-02
#>  [272,]  0.19657240  7.073233e-02 -6.882941e-03
#>  [273,]  0.22377522  9.611311e-02 -1.974975e-02
#>  [274,]  0.19592515  8.557019e-02 -6.488568e-03
#>  [275,]  0.22646786  9.129908e-02 -1.139453e-02
#>  [276,]  0.19601449  6.811735e-02  1.281790e-02
#>  [277,]  0.22277795  8.549428e-02 -1.015566e-02
#>  [278,]  0.20964339  9.758093e-02 -1.920746e-02
#>  [279,]  0.22279961  1.115858e-01  1.503323e-02
#>  [280,]  0.21240691  8.442615e-02 -1.677795e-02
#>  [281,]  0.21753451  8.963746e-02  3.864444e-03
#>  [282,]  0.20389441  9.281011e-02 -4.927002e-03
#>  [283,]  0.21338769  8.461239e-02 -1.669194e-02
#>  [284,]  0.20908042  6.947005e-02 -6.490885e-03
#>  [285,]  0.22601754  9.231383e-02 -1.396417e-02
#>  [286,]  0.17196663  6.159564e-02 -2.812701e-02
#>  [287,]  0.19819501  6.521188e-02 -2.454225e-02
#>  [288,]  0.19823828  9.754204e-02 -4.186544e-03
#>  [289,]  0.17481407  6.232667e-02 -6.329869e-03
#>  [290,]  0.19209221  5.938876e-02 -2.570847e-02
#>  [291,]  0.22378408 -1.349556e-03  2.827464e-02
#>  [292,]  0.21318605  7.050277e-02 -1.555423e-02
#>  [293,]  0.20922760  2.096058e-02  4.110684e-02
#>  [294,]  0.21753894  7.389570e-02 -7.341658e-03
#>  [295,]  0.24171902  9.957953e-02 -2.236560e-02
#>  [296,]  0.19786782  7.153989e-02 -4.923070e-03
#>  [297,]  0.21585727  8.818371e-02 -8.872206e-04
#>  [298,]  0.21396143  8.043316e-02 -2.883633e-03
#>  [299,]  0.22451194  8.098156e-02  1.880392e-02
#>  [300,]  0.18664824  4.476375e-02  2.605052e-02
#>  [301,]  0.22369247  9.303572e-02 -1.339756e-02
#>  [302,]  0.18350031  6.252620e-02 -2.363976e-03
#>  [303,]  0.20466147  7.200467e-02 -1.389776e-02
#>  [304,]  0.22105408  9.453270e-02 -8.719805e-03
#>  [305,]  0.16862757  7.054928e-02  7.389111e-03
#>  [306,]  0.23460142  9.568338e-02 -6.143568e-03
#>  [307,]  0.17843706  5.148999e-02 -2.142666e-02
#>  [308,]  0.23580285  8.644305e-02 -2.131586e-02
#>  [309,]  0.23601693  7.376624e-02 -1.744635e-02
#>  [310,]  0.22546907  8.963314e-02 -5.605740e-04
#>  [311,]  0.20619911  5.931303e-02 -8.026225e-03
#>  [312,]  0.21619770  7.857594e-02 -1.443905e-02
#>  [313,]  0.22273018  9.619224e-02 -1.594501e-02
#>  [314,]  0.21893045  8.252204e-02  1.964856e-03
#>  [315,]  0.22094927  7.625301e-02 -2.515151e-02
#>  [316,]  0.21882958  9.203176e-02 -6.995706e-03
#>  [317,]  0.17969015  4.322492e-02 -2.027980e-02
#>  [318,]  0.24453737  1.051315e-01 -8.774366e-03
#>  [319,]  0.24311259  1.032893e-01 -6.125767e-03
#>  [320,]  0.20448575  7.862528e-02 -9.726472e-03
#>  [321,]  0.18162953  9.357403e-03 -1.353067e-02
#>  [322,]  0.18577366  5.534008e-02 -2.797664e-02
#>  [323,]  0.20538109  6.939733e-02 -1.610821e-02
#>  [324,]  0.23161954  1.025527e-01 -8.671050e-02
#>  [325,]  0.15651288  6.418301e-02 -8.290881e-02
#>  [326,]  0.20397346  7.652500e-02 -1.240126e-02
#>  [327,]  0.21652576  1.165157e-01 -1.850514e-02
#>  [328,]  0.20858572  7.403126e-02 -2.348363e-03
#>  [329,]  0.21271442  7.611795e-02 -1.272847e-02
#>  [330,]  0.20923828  9.581869e-02 -1.457558e-02
#>  [331,]  0.22961087  9.059282e-02 -3.475637e-03
#>  [332,]  0.18178252  5.576484e-02 -1.609237e-02
#>  [333,]  0.20289686  7.285856e-02 -1.511759e-02
#>  [334,]  0.21538168  8.907218e-02  2.682803e-04
#>  [335,]  0.21038739  4.082043e-02  4.706009e-03
#>  [336,]  0.17479707  1.213703e-01 -2.716243e-02
#>  [337,]  0.22161479  8.904895e-02 -1.430800e-02
#>  [338,]  0.18877917  6.908134e-02 -1.663239e-02
#>  [339,]  0.26722654  8.983072e-02 -1.862677e-02
#>  [340,]  0.20921282  8.180556e-02 -1.144439e-02
#>  [341,]  0.20809821  1.015448e-01 -2.888517e-02
#>  [342,]  0.19396082  7.332254e-02  5.770857e-03
#>  [343,]  0.22513328  9.973533e-02 -2.033042e-02
#>  [344,]  0.29286822  7.068284e-02 -2.669580e-03
#>  [345,]  0.19848808  5.650204e-02 -1.940650e-02
#>  [346,]  0.21709764  8.801116e-02 -2.923782e-02
#>  [347,]  0.21829078  8.876132e-02 -2.031523e-02
#>  [348,]  0.16901269  5.098700e-02 -3.061428e-02
#>  [349,]  0.21179630  9.117997e-02 -1.243839e-02
#>  [350,]  0.22077292  7.557868e-02 -1.901188e-02
#>  [351,]  0.18075510  5.806318e-02 -2.373099e-02
#>  [352,]  0.19163291  4.715955e-02 -2.288051e-02
#>  [353,]  0.20742838  8.415121e-02 -7.996668e-03
#>  [354,]  0.19778044  7.157110e-02 -1.973791e-02
#>  [355,]  0.19651464  3.990904e-02 -2.681857e-03
#>  [356,]  0.19393210  6.888440e-02 -1.194160e-02
#>  [357,]  0.22933127  1.050096e-01  3.888280e-03
#>  [358,]  0.21428528  9.715484e-02 -2.189056e-02
#>  [359,]  0.22209705  9.437280e-02 -9.628452e-03
#>  [360,]  0.19196983  6.722809e-02 -1.806729e-02
#>  [361,]  0.01516548  5.986567e-02 -5.378609e-03
#>  [362,]  0.25188389  7.975435e-02 -2.017314e-02
#>  [363,]  0.22569044  5.401005e-02 -1.344759e-02
#>  [364,]  0.17778750  5.063049e-02 -1.219075e-02
#>  [365,]  0.21394131  8.266726e-02  4.439781e-04
#>  [366,]  0.22308495  7.888203e-02  1.806928e-02
#>  [367,]  0.24041073  1.019462e-01 -1.261197e-02
#>  [368,]  0.17525321  6.901431e-02  2.627950e-03
#>  [369,]  0.22449102  8.727701e-02 -1.475293e-02
#>  [370,]  0.20117776  8.124021e-02  3.636758e-03
#>  [371,]  0.22644187  1.079957e-01 -4.316152e-02
#>  [372,]  0.22714139 -6.745355e-02 -1.584046e-02
#>  [373,]  0.21812976  9.234375e-02 -3.109341e-03
#>  [374,]  0.20054766  8.163449e-02 -9.356302e-03
#>  [375,]  0.22943844  2.340725e-02  5.063360e-03
#>  [376,]  0.19375573  5.912572e-02 -4.627476e-02
#>  [377,]  0.22357723  9.411850e-02 -1.119550e-02
#>  [378,]  0.22471838  7.853623e-02 -1.649356e-02
#>  [379,]  0.17085089  5.703574e-02 -8.095527e-03
#>  [380,]  0.20497569  7.894534e-02 -1.689791e-02
#>  [381,]  0.20949446  9.489973e-02  2.302607e-02
#>  [382,]  0.22247850  9.144958e-02  8.906678e-03
#>  [383,]  0.23041508  9.889692e-02 -1.067418e-02
#>  [384,]  0.21858420  8.167789e-02  1.363459e-03
#>  [385,]  0.23174315  1.182522e-01 -4.664376e-03
#>  [386,]  0.23704196  9.635526e-02 -7.625102e-03
#>  [387,]  0.23188576  9.653349e-02 -2.137036e-02
#>  [388,]  0.17844750  4.604381e-02 -2.196924e-02
#>  [389,]  0.16853101  5.204617e-02 -3.462931e-02
#>  [390,]  0.22460631  8.612171e-02 -7.828930e-03
#>  [391,]  0.21293337  8.189999e-02 -4.182681e-03
#>  [392,]  0.22000079  8.641064e-02 -1.338992e-02
#>  [393,]  0.19652428  7.180854e-02 -1.172067e-02
#>  [394,]  0.21004083  8.727109e-02 -1.470745e-02
#>  [395,]  0.19270968  6.409970e-02  2.864855e-03
#>  [396,]  0.21413186  8.316620e-02 -8.896768e-03
#>  [397,]  0.18288848  5.919907e-02 -9.905938e-03
#>  [398,]  0.20995876  8.979374e-02  4.742552e-03
#>  [399,]  0.22209483  9.794661e-02 -2.940885e-02
#>  [400,]  0.23005037  9.722182e-02 -1.940594e-02
#>  [401,]  0.23505274  5.660470e-02 -2.112591e-02
#>  [402,]  0.22466873  9.120466e-02 -4.506957e-03
#>  [403,]  0.20707309  8.735580e-02 -1.604403e-03
#>  [404,]  0.22098677  9.059757e-02 -9.974138e-03
#>  [405,]  0.16842837  6.288278e-02 -2.243862e-03
#>  [406,]  0.19360957  5.227642e-02 -1.562822e-02
#>  [407,]  0.28675014  1.049248e-01 -2.052735e-02
#>  [408,]  0.20480887  6.855516e-02 -2.038256e-02
#>  [409,]  0.22552599  9.404003e-02 -7.304575e-03
#>  [410,]  0.17357846  2.087282e-01 -8.311318e-03
#>  [411,]  0.16028963  2.105815e-01 -9.766224e-04
#>  [412,]  0.20629581  7.037631e-02 -1.806203e-02
#>  [413,]  0.20799305  9.291919e-02  1.138143e-02
#>  [414,]  0.22788964  9.575146e-02 -1.446905e-02
#>  [415,]  0.21506008  6.925653e-02  2.033512e-03
#>  [416,]  0.24205211  5.780879e-02  3.087340e-02
#>  [417,]  0.21641139  8.524733e-02 -1.437053e-02
#>  [418,]  0.27765990  1.336728e-01 -1.594884e-02
#>  [419,]  0.20101532  7.730101e-02 -1.434380e-02
#>  [420,]  0.20509142  5.079702e-02  1.411792e-02
#>  [421,]  0.21528198  3.208999e-02 -5.040093e-02
#>  [422,]  0.19616458  9.062026e-02 -1.569781e-02
#>  [423,]  0.18227606  7.566184e-02 -7.180088e-03
#>  [424,]  0.23113464  9.450387e-02 -7.303034e-03
#>  [425,]  0.19895822  6.220705e-02 -1.746282e-02
#>  [426,]  0.22284824  9.200039e-02 -1.129518e-02
#>  [427,]  0.24526016  1.032299e-01 -4.909897e-03
#>  [428,]  0.18458667  7.399718e-02 -2.086962e-02
#>  [429,]  0.21696727  7.161435e-02 -2.093771e-02
#>  [430,]  0.24931630  1.064145e-01 -9.867946e-03
#>  [431,]  0.20962443  8.401637e-02 -2.132871e-02
#>  [432,]  0.21128860  3.714747e-02  9.272920e-03
#>  [433,]  0.21804041  1.035906e-01 -1.463803e-02
#>  [434,]  0.20564416  7.125553e-02 -8.959180e-03
#>  [435,]  0.20667750  7.857552e-02 -4.151019e-03
#>  [436,]  0.23094295  9.850548e-02 -2.886170e-03
#>  [437,]  0.21925396  8.874203e-02 -1.960500e-02
#>  [438,]  0.21397334  8.281558e-02 -1.309705e-02
#>  [439,]  0.21559238  8.105665e-02 -7.732429e-03
#>  [440,]  0.23099663  1.044671e-01 -5.748882e-02
#>  [441,]  0.19098528  6.333877e-02 -1.912880e-02
#>  [442,]  0.20483285  4.729349e-02 -3.459364e-03
#>  [443,]  0.18170680  8.812864e-02 -3.080860e-02
#>  [444,]  0.21198200  8.687613e-02 -1.454125e-02
#>  [445,]  0.21181064  9.016326e-02  7.072854e-04
#>  [446,]  0.22130557  6.638529e-02 -8.323286e-03
#>  [447,]  0.20695874  8.377363e-02 -2.011236e-02
#>  [448,]  0.22418842 -1.221815e-01  1.697412e-01
#>  [449,]  0.19844137  6.838342e-02 -1.583229e-02
#>  [450,]  0.20203516  8.034509e-02  1.763136e-03
#>  [451,]  0.24039100  1.021855e-01 -9.753579e-03
#>  [452,]  0.22978930  9.892488e-02 -1.608198e-02
#>  [453,]  0.21440195  8.321887e-02 -2.043425e-02
#>  [454,]  0.21449536  8.262410e-02 -7.385299e-03
#>  [455,]  0.16027627  8.685548e-02  2.424436e-02
#>  [456,]  0.18897513  1.926635e-01  1.984101e-02
#>  [457,]  0.20459103  6.297611e-02 -1.349762e-02
#>  [458,]  0.20438195  7.502831e-02 -2.869095e-02
#>  [459,]  0.18274814  8.172351e-02 -6.568941e-03
#>  [460,]  0.21852134  8.824670e-02 -1.837940e-02
#>  [461,]  0.20853797  7.262687e-02 -1.443302e-02
#>  [462,]  0.19961190  6.502619e-02 -1.559301e-02
#>  [463,]  0.19955106  1.003551e-01  8.370359e-03
#>  [464,]  0.21728252  7.172813e-02 -3.030979e-02
#>  [465,]  0.21700842  7.878588e-02 -7.179393e-03
#>  [466,]  0.21801568  9.435870e-02 -1.598522e-02
#>  [467,]  0.21866578  7.495244e-02 -3.106133e-02
#>  [468,]  0.15094844  1.921526e-01 -5.688352e-03
#>  [469,]  0.19592203  8.358720e-02  4.931808e-03
#>  [470,]  0.16808375  6.387442e-02  2.456895e-02
#>  [471,]  0.22981717  9.535235e-02 -2.180947e-03
#>  [472,]  0.24024661  9.778204e-02 -1.076501e-02
#>  [473,]  0.22024351  8.427285e-02 -1.175852e-02
#>  [474,]  0.19993211  1.011306e-01 -9.237454e-03
#>  [475,]  0.18740845  6.918474e-02 -6.865272e-02
#>  [476,]  0.20472359  7.808686e-02 -8.251145e-03
#>  [477,]  0.21153508  9.624609e-02  1.434190e-02
#>  [478,]  0.23473474  9.531767e-02 -2.014590e-02
#>  [479,]  0.20972014  7.738566e-02 -2.095114e-02
#>  [480,]  0.22044886  9.416775e-02 -9.014880e-03
#>  [481,]  0.22117213  9.286675e-02 -1.769219e-02
#>  [482,]  0.18357414  6.734067e-02 -3.285987e-02
#>  [483,]  0.21833093  8.349180e-02 -9.659157e-03
#>  [484,]  0.17323326  3.143426e-02  9.522827e-02
#>  [485,]  0.21394811  8.905322e-02  9.105750e-03
#>  [486,]  0.25926312  9.619880e-02 -1.063681e-02
#>  [487,]  0.19104347  6.597646e-02 -1.062090e-02
#>  [488,]  0.21522412  8.792778e-02 -8.944075e-03
#>  [489,]  0.19686521  6.029902e-02 -1.183455e-02
#>  [490,]  0.22528193  1.069110e-01 -1.978750e-02
#>  [491,]  0.19090618  8.169463e-02 -9.398708e-03
#>  [492,]  0.21860240  9.150167e-02 -1.979489e-02
#>  [493,]  0.22098780  8.641613e-02 -5.391765e-03
#>  [494,]  0.23637871  9.981369e-02 -3.479409e-02
#>  [495,]  0.20621625  1.113552e-01 -2.007549e-03
#>  [496,]  0.18755671  6.106711e-02 -5.253797e-02
#>  [497,]  0.21417335  7.892994e-02 -2.563573e-02
#>  [498,]  0.20967100  8.916935e-02 -9.192376e-02
#>  [499,]  0.19259298  7.544888e-02 -1.792479e-02
#>  [500,]  0.19229332  6.339566e-02 -1.798184e-02
#>  [501,]  0.22541523  8.914839e-02 -1.590722e-02
#>  [502,]  0.20774909  8.265249e-02 -1.559004e-02
#>  [503,]  0.24813896  1.362255e-01  3.934212e-02
#>  [504,]  0.24171193  9.379191e-02 -1.469862e-02
#>  [505,]  0.19309669  7.711276e-02 -3.532882e-02
#>  [506,]  0.21305064  1.002851e-01  1.638706e-03
#>  [507,]  0.19937316  6.459233e-02 -1.939810e-02
#>  [508,]  0.09056865 -2.823212e-02 -1.126562e-01
#>  [509,]  0.19158563  8.150995e-02 -7.897844e-02
#>  [510,]  0.18822570  5.812249e-02 -2.558620e-02
#>  [511,]  0.17486107  4.249748e-02 -4.197641e-03
#>  [512,]  0.20253338  5.625710e-02 -1.202124e-02
#>  [513,]  0.18513784  6.361616e-02 -1.364567e-02
#>  [514,]  0.21304180  7.819948e-02 -6.393606e-03
#>  [515,]  0.21592432  8.938134e-02 -1.292879e-02
#>  [516,]  0.23418338  7.598942e-02 -3.953736e-02
#>  [517,]  0.23308111  8.937797e-02 -1.581766e-02
#>  [518,]  0.17811637  8.831550e-02 -1.504343e-02
#>  [519,]  0.19909474  6.868909e-02 -1.896603e-03
#>  [520,]  0.22459646  7.131545e-02 -1.686597e-02
#>  [521,]  0.30124523  1.089965e-01 -3.757317e-02
#>  [522,]  0.24424157  1.040484e-01 -8.062498e-03
#>  [523,]  0.20761019  1.005411e-01 -3.081599e-02
#>  [524,]  0.12483850 -3.084068e-02 -1.101210e-01
#>  [525,]  0.19095963  6.328860e-02  6.387948e-03
#>  [526,]  0.19962137  7.907268e-02 -2.381556e-02
#>  [527,]  0.21734473  8.876353e-02 -1.142775e-02
#>  [528,]  0.20179877  7.560965e-02 -1.016686e-02
#>  [529,]  0.22903613  9.729037e-02 -9.688103e-03
#>  [530,]  0.17842085  5.482398e-02 -2.078376e-02
#>  [531,]  0.21071478  7.835077e-02 -6.843020e-03
#>  [532,]  0.19189408  7.143177e-02 -7.024059e-03
#>  [533,]  0.19155996  5.229015e-02 -1.379932e-02
#>  [534,]  0.20682689  8.074700e-02 -8.698003e-03
#>  [535,]  0.21566050  8.112270e-02 -1.608149e-02
#>  [536,]  0.23150697  8.939470e-02 -1.405044e-02
#>  [537,]  0.20831983  8.190254e-02 -1.518999e-02
#>  [538,]  0.21854604  8.062137e-02 -1.942528e-02
#>  [539,]  0.16178267  9.995422e-02 -6.914747e-02
#>  [540,]  0.37965621  1.050155e-01 -2.238145e-02
#>  [541,]  0.03555448  8.534570e-02 -1.228892e-02
#>  [542,]  0.20426690  8.575978e-02  1.477750e-03
#>  [543,]  0.23484765  8.281409e-02 -1.095288e-02
#>  [544,]  0.24532771  8.763389e-02 -1.719088e-02
#>  [545,]  0.19968769  8.090067e-02 -2.499082e-02
#>  [546,]  0.23051981  9.384254e-02 -1.066932e-02
#>  [547,]  0.18856103  5.268855e-02  3.879631e-03
#>  [548,]  0.18908472  6.611876e-02 -1.145326e-02
#>  [549,]  0.19455189  1.099388e-01 -8.252351e-03
#>  [550,]  0.21547214  9.260447e-02  4.343345e-02
#>  [551,]  0.20755637  7.950428e-02 -2.244575e-02
#>  [552,]  0.23144819  9.307972e-02 -2.542074e-03
#>  [553,]  0.16486615  5.994882e-02 -1.736021e-02
#>  [554,]  0.22125550  8.847524e-02 -4.338149e-03
#>  [555,]  0.22150409  9.092244e-02 -8.795868e-03
#>  [556,]  0.23754677  1.013022e-01 -8.780308e-03
#>  [557,]  0.19558569  4.026541e-02  3.443013e-02
#>  [558,]  0.19619898  7.379177e-02 -2.744480e-02
#>  [559,]  0.20512826  7.843028e-02 -3.893684e-03
#>  [560,]  0.16761316  5.659506e-02 -1.264678e-02
#>  [561,]  0.23275805  1.017622e-01 -7.210861e-03
#>  [562,]  0.19857571  8.134067e-02 -5.951310e-03
#>  [563,]  0.20764441  2.690872e-02 -1.103213e-02
#>  [564,]  0.22244379  8.402609e-02 -1.180596e-02
#>  [565,]  0.21060100  5.142765e-02 -1.756106e-02
#>  [566,]  0.19828478  7.597396e-02 -1.965803e-02
#>  [567,]  0.22203903  8.569532e-02 -7.939952e-03
#>  [568,]  0.18811461  7.173040e-02 -7.764842e-04
#>  [569,]  0.23031375  9.147884e-02 -1.530543e-02
#>  [570,]  0.12615528  3.927164e-02 -1.577316e-02
#>  [571,]  0.22459096  8.788850e-02 -1.437074e-02
#>  [572,]  0.20579363  3.887536e-02 -1.937018e-02
#>  [573,]  0.21605454  1.136760e-01  2.913692e-04
#>  [574,]  0.22016550  9.105941e-02 -8.143502e-03
#>  [575,]  0.20485694  8.596790e-02 -9.878996e-03
#>  [576,]  0.19527024  7.351206e-02 -1.063260e-02
#>  [577,]  0.20101971  6.923074e-02 -7.608341e-03
#>  [578,]  0.21531901  7.890924e-02 -3.325171e-03
#>  [579,]  0.18105381  6.732008e-02 -2.150668e-02
#>  [580,]  0.23579410  8.056325e-02 -5.087862e-02
#>  [581,]  0.22540888  7.877935e-02 -1.228586e-02
#>  [582,]  0.19879593  6.793668e-02 -1.144650e-02
#>  [583,]  0.21721071  6.783302e-02 -6.601090e-02
#>  [584,]  0.22575355  7.752724e-02 -1.004914e-02
#>  [585,]  0.21772492  7.402134e-02 -4.129959e-03
#>  [586,]  0.20165041  7.222315e-02 -1.766387e-02
#>  [587,]  0.21571552  8.105856e-02 -6.665701e-03
#>  [588,]  0.22160520  9.660593e-02 -3.249839e-03
#>  [589,]  0.23265899  9.754264e-02 -1.310344e-02
#>  [590,]  0.19594410  6.841866e-02 -1.580815e-02
#>  [591,]  0.19432096  8.480463e-02 -1.153414e-02
#>  [592,]  0.19421690  7.717826e-02 -2.230139e-02
#>  [593,]  0.19256390  7.287285e-02 -2.254810e-03
#>  [594,]  0.21823666  8.734927e-02 -1.208281e-02
#>  [595,]  0.20928066  7.541526e-02 -4.057370e-03
#>  [596,]  0.23270934  9.937052e-02 -1.328959e-02
#>  [597,]  0.19850414  4.285895e-02 -1.762844e-02
#>  [598,]  0.20118574  5.160613e-02 -5.410302e-02
#>  [599,]  0.16200111  7.561554e-02 -3.619053e-03
#>  [600,]  0.21719036  9.434377e-02 -1.419059e-02
#>  [601,]  0.20731653  9.239691e-02 -1.274482e-02
#>  [602,]  0.18199016  7.007318e-02 -1.694265e-02
#>  [603,]  0.20510089  7.731590e-02 -1.759465e-02
#>  [604,]  0.21311801  8.006848e-02 -2.243215e-02
#>  [605,]  0.21970487  8.722200e-02 -1.386289e-02
#>  [606,]  0.22464499  8.842212e-02 -2.065521e-03
#>  [607,]  0.19290259  7.953602e-02 -2.688605e-02
#>  [608,]  0.21351972  9.182437e-02 -5.212982e-03
#>  [609,]  0.23669573  9.805238e-02 -2.158465e-02
#>  [610,]  0.17171811  1.211736e-01 -2.900035e-02
#>  [611,]  0.21068298  7.180922e-02 -1.617808e-02
#>  [612,]  0.21106001  8.533953e-02 -6.426020e-03
#>  [613,]  0.20577939  8.620332e-02 -2.855649e-02
#>  [614,]  0.22531257  9.363691e-02 -1.541746e-02
#>  [615,]  0.20247575  8.379569e-02 -1.447557e-02
#>  [616,]  0.22788976  1.333196e-01  1.480406e-02
#>  [617,]  0.21796181  8.324639e-02 -1.410618e-02
#>  [618,]  0.23353785  9.609856e-02 -1.181737e-02
#>  [619,]  0.21286984  9.776310e-02 -2.222720e-02
#>  [620,]  0.20812809  7.496402e-02 -1.000260e-02
#>  [621,]  0.21209600  8.339979e-02 -1.210633e-02
#>  [622,]  0.21933013  8.362718e-02 -9.584511e-03
#>  [623,]  0.20653729  7.130071e-02 -1.685349e-02
#>  [624,]  0.19612711  6.810459e-02  6.117557e-03
#>  [625,]  0.19767734  9.597002e-02  4.383217e-03
#>  [626,]  0.20157908  7.260082e-02 -1.225266e-02
#>  [627,]  0.22032181  7.219256e-02  2.491697e-04
#>  [628,]  0.20826871  7.604294e-02  2.305268e-02
#>  [629,]  0.18504856  5.399883e-02 -2.140326e-02
#>  [630,]  0.21177569  2.029978e-01  5.868316e-02
#>  [631,]  0.28556826 -4.984887e-02 -5.625104e-02
#>  [632,]  0.22209729  1.008407e-01 -1.652449e-02
#>  [633,]  0.14916906  4.309838e-02 -3.806127e-02
#>  [634,]  0.20165471  8.685214e-02 -3.911100e-03
#>  [635,]  0.20878371  7.383307e-02 -1.284859e-02
#>  [636,]  0.21649400  9.330629e-02 -2.735772e-02
#>  [637,]  0.03274588 -5.877423e-02  3.945825e-01
#>  [638,]  0.21235801  9.044982e-02 -1.583950e-02
#>  [639,]  0.20084605  8.741621e-02  3.027591e-02
#>  [640,]  0.20463364  7.560375e-02 -2.362956e-02
#>  [641,]  0.23192675  7.580412e-02 -1.133459e-02
#>  [642,]  0.22190011  8.834015e-02 -1.388946e-02
#>  [643,]  0.09492495  1.169624e-02 -3.856949e-02
#>  [644,]  0.20630579  8.689770e-02 -5.447440e-03
#>  [645,]  0.21015138  1.043010e-01 -5.440014e-03
#>  [646,]  0.18448970  7.747070e-02 -6.505510e-02
#>  [647,]  0.21461443  6.237090e-02 -4.826647e-02
#>  [648,]  0.22404439  9.099774e-02 -1.310738e-02
#>  [649,]  0.22242822  8.672761e-02  1.486489e-03
#>  [650,]  0.24288362  1.021527e-01 -1.322343e-02
#>  [651,]  0.21805026  8.756875e-02 -1.011516e-02
#>  [652,]  0.16613125  5.059854e-02 -6.266765e-02
#>  [653,]  0.24000180  9.337317e-02 -1.446918e-02
#>  [654,]  0.22411155  5.241192e-02  9.205296e-02
#>  [655,]  0.21631022  8.027768e-02 -5.498628e-03
#>  [656,]  0.21120768  7.877148e-02 -1.520687e-02
#>  [657,]  0.20306281  7.526494e-02 -1.277761e-02
#>  [658,]  0.22305275  8.702999e-02 -2.014965e-02
#>  [659,]  0.20613139  7.898975e-02 -9.501959e-03
#>  [660,]  0.20286380  9.011894e-02 -9.024886e-04
#>  [661,]  0.19197261  6.795733e-02  1.421749e-02
#>  [662,]  0.20984925  1.037052e-01 -3.037549e-02
#>  [663,]  0.20328406  7.392272e-02 -3.621079e-03
#>  [664,]  0.20540168  4.091660e-02  1.967521e-02
#>  [665,]  0.19379396  6.896209e-02 -1.405546e-02
#>  [666,]  0.24459424  1.041359e-01 -5.627448e-02
#>  [667,]  0.22850782  9.533804e-02 -3.335964e-02
#>  [668,]  0.20599082  9.669761e-02 -9.698160e-03
#>  [669,]  0.18312450  6.331259e-02 -2.152624e-02
#>  [670,]  0.23416278  7.367898e-02 -1.216355e-02
#>  [671,]  0.19116746  6.187568e-02 -1.828367e-02
#>  [672,]  0.23060022  9.283408e-02 -1.655252e-02
#>  [673,]  0.20630454  5.428068e-02 -1.328490e-02
#>  [674,]  0.21072660  9.450855e-02 -8.263669e-03
#>  [675,]  0.23504389  1.063045e-01 -8.121254e-04
#>  [676,]  0.19688515  2.818283e-02 -2.853604e-02
#>  [677,]  0.20980519  7.696164e-02 -6.543640e-03
#>  [678,]  0.18120814  1.750576e-01  2.619250e-02
#>  [679,]  0.21761283  1.188106e-01 -3.007253e-02
#>  [680,]  0.23993970  8.238998e-02 -2.576504e-02
#>  [681,]  0.21016169  8.592604e-02 -2.215822e-02
#>  [682,]  0.24359624  9.956520e-02 -2.192674e-04
#>  [683,]  0.22073233  8.215346e-02  7.102665e-02
#>  [684,]  0.19971954  1.247559e-01  3.485594e-02
#>  [685,]  0.21215267  7.978259e-02 -1.044114e-02
#>  [686,]  0.20656114  8.026827e-02 -1.652001e-02
#>  [687,]  0.20370213  7.588590e-02  1.126298e-02
#>  [688,]  0.19917504  7.357396e-02 -1.806081e-02
#>  [689,]  0.20752178  8.520643e-02 -8.602174e-03
#>  [690,]  0.22999322  7.669286e-02 -4.865052e-03
#>  [691,]  0.20234790  6.923601e-02 -1.710395e-02
#>  [692,]  0.23073366  1.038550e-01 -1.097944e-02
#>  [693,]  0.20320301  6.837472e-02  4.646136e-03
#>  [694,]  0.21499222  2.435918e-02  1.073935e-01
#>  [695,]  0.21784070  9.314278e-02 -2.601975e-02
#>  [696,]  0.20030815  1.140480e-01  1.352748e-03
#>  [697,]  0.22228043  8.901454e-02 -1.165531e-02
#>  [698,]  0.17989543  6.935426e-02 -1.753157e-02
#>  [699,]  0.21877391  7.984477e-02 -1.943935e-02
#>  [700,]  0.18363447  6.756018e-02 -8.371846e-03
#>  [701,]  0.24408726  1.059584e-01 -1.066618e-02
#>  [702,]  0.20231679  7.527100e-02 -6.378177e-03
#>  [703,]  0.23034258  8.736006e-02 -1.854350e-02
#>  [704,]  0.24133540  1.076664e-01 -4.125812e-03
#>  [705,]  0.17124099  5.241033e-02 -7.044955e-02
#>  [706,]  0.19962902  7.660377e-02 -1.542551e-02
#>  [707,]  0.22070608  4.570501e-02 -2.068740e-02
#>  [708,]  0.23291462  9.481482e-02 -1.588255e-02
#>  [709,]  0.20184633  8.641636e-02 -2.105570e-02
#>  [710,]  0.20875929  7.094092e-02 -4.813609e-03
#>  [711,]  0.18047764  4.968500e-02 -1.988333e-02
#>  [712,]  0.21575773  1.002838e-01 -8.772509e-03
#>  [713,]  0.19858631  5.093460e-02 -1.538017e-03
#>  [714,]  0.21402219  8.078294e-02 -1.427874e-02
#>  [715,]  0.20668257  7.583425e-02  3.500955e-03
#>  [716,]  0.21911584  8.734869e-02 -1.192183e-02
#>  [717,]  0.21884142  5.009980e-02 -1.209100e-02
#>  [718,]  0.20744405  5.552407e-02 -5.077656e-02
#>  [719,]  0.22565705  1.081505e-01  2.173091e-02
#>  [720,]  0.20505983  5.186506e-02 -3.193789e-02
#>  [721,]  0.17377746  5.572457e-02 -1.693595e-03
#>  [722,]  0.21024173  8.848522e-02 -1.697539e-02
#>  [723,]  0.23175005  9.275410e-02 -1.532935e-02
#>  [724,]  0.24284587  1.031171e-01 -8.573219e-03
#>  [725,]  0.20077504  7.549434e-02 -1.116610e-02
#>  [726,]  0.18841564  1.366031e-01 -2.736020e-02
#>  [727,]  0.18128290  5.974586e-02 -3.302516e-02
#>  [728,]  0.21234115  7.897998e-02 -1.752087e-02
#>  [729,]  0.12656491  9.914666e-02  2.557350e-02
#>  [730,]  0.20775099  6.562924e-02 -2.730082e-02
#>  [731,]  0.21685796  8.927978e-02 -2.466498e-02
#>  [732,]  0.20030298  4.249079e-02 -2.351984e-02
#>  [733,]  0.21201071  7.324826e-02 -1.043094e-02
#>  [734,]  0.18344041  7.098417e-02 -6.289556e-03
#>  [735,]  0.19671371  7.045129e-02  5.752405e-02
#>  [736,]  0.18014077  5.493441e-02 -1.068606e-02
#>  [737,]  0.19772899  7.496852e-02  9.349009e-03
#>  [738,]  0.21109274  8.922738e-02  2.105290e-02
#>  [739,]  0.19589847  9.072946e-02  8.193108e-02
#>  [740,]  0.22939733  9.807737e-02 -8.762536e-03
#>  [741,]  0.21120776  1.893822e-01  6.901577e-03
#>  [742,]  0.20005710  8.480518e-02 -6.864397e-02
#>  [743,]  0.20296909  1.006668e-01 -1.437313e-01
#>  [744,]  0.21335320  8.124684e-02 -1.582337e-02
#>  [745,]  0.21249208  8.441643e-02 -1.393319e-02
#>  [746,]  0.19817275  7.229260e-02 -1.147961e-02
#>  [747,]  0.20879364  7.991583e-02 -2.751929e-02
#>  [748,]  0.13247637  5.328493e-02  3.486473e-02
#>  [749,]  0.23620966  8.724768e-02 -1.629326e-02
#>  [750,]  0.21078607  8.356212e-02 -8.553128e-03
#>  [751,]  0.19338472  6.676497e-02 -9.173154e-03
#>  [752,]  0.19775442  8.318540e-02 -2.853353e-03
#>  [753,]  0.16652810  5.561249e-02 -1.629694e-02
#>  [754,]  0.23281121  9.386862e-02 -1.502969e-02
#>  [755,]  0.21433949  8.799924e-02 -1.311743e-02
#>  [756,]  0.22148485  9.043865e-02 -1.102282e-02
#>  [757,] -0.49170157 -1.234535e-01  5.068217e-03
#>  [758,]  0.21067468  5.417779e-02 -2.398819e-02
#>  [759,]  0.21779935  8.684919e-02 -9.802493e-03
#>  [760,]  0.21513168  8.138595e-02 -1.252148e-02
#>  [761,]  0.22404806  9.241823e-02 -9.654269e-03
#>  [762,]  0.18563484  6.196552e-02 -6.727866e-03
#>  [763,]  0.21727761  7.482207e-02 -2.499353e-02
#>  [764,]  0.22424424  8.414169e-02 -2.170541e-02
#>  [765,]  0.18172143  6.729844e-02 -2.422041e-02
#>  [766,]  0.21789405  8.251040e-02 -1.589603e-02
#>  [767,]  0.22813180  7.746832e-02  5.441298e-02
#>  [768,]  0.18575920  1.142836e-01 -1.842913e-02
#>  [769,]  0.22323441  8.936432e-02 -1.127262e-02
#>  [770,]  0.24050348  1.011747e-01 -9.651733e-03
#>  [771,]  0.19603593  6.642910e-02 -1.918184e-02
#>  [772,]  0.20339242  8.920431e-02 -3.935928e-02
#>  [773,]  0.23295325  9.782423e-02 -6.576858e-03
#>  [774,]  0.15987237 -6.547188e-05 -1.717542e-01
#>  [775,]  0.12661831  2.738779e-02 -4.223764e-02
#>  [776,]  0.21655970  8.365162e-02 -1.053680e-02
#>  [777,]  0.21457368  2.745363e-02 -2.567529e-02
#>  [778,]  0.22684387  9.177066e-02 -1.149005e-02
#>  [779,]  0.27915646  2.014386e-01  9.913676e-02
#>  [780,]  0.19108112  8.843311e-02 -3.877612e-02
#>  [781,]  0.18307985  6.415130e-02 -3.139037e-02
#>  [782,]  0.18391157  6.848156e-02 -9.902129e-03
#>  [783,]  0.18496095  1.258553e-01  9.363999e-03
#>  [784,]  0.23642313  9.896338e-02 -6.112876e-02
#>  [785,]  0.23282524  1.035807e-01 -6.584510e-03
#>  [786,]  0.21143981  7.988119e-02 -9.271401e-03
#>  [787,]  0.22364224  9.082441e-02 -1.113447e-02
#>  [788,]  0.19900705  6.942064e-02 -2.295840e-02
#>  [789,]  0.17672836  3.712493e-02 -4.765254e-02
#>  [790,] -0.04704040 -2.141423e-01 -2.679050e-01
#>  [791,]  0.22541838  9.500559e-02 -8.597319e-03
#>  [792,]  0.19827938  7.479791e-02 -2.125872e-02
#>  [793,]  0.22369219  1.797686e-02 -2.688756e-03
#>  [794,]  0.08694513  1.969154e-01 -3.945825e-01
#>  [795,]  0.21463199  8.426007e-02 -1.137903e-02
#>  [796,]  0.19001162  8.382637e-02 -1.122535e-02
#>  [797,]  0.22279036  8.762921e-02 -1.191705e-02
#>  [798,]  0.22829102  6.466536e-02 -2.547727e-02
#>  [799,]  0.15058830  7.162077e-02  2.828653e-02
#>  [800,]  0.19613157  1.611942e-01  6.837389e-02
#>  [801,]  0.20843881  7.643773e-02 -2.087458e-02
#>  [802,]  0.20654486  8.116152e-02 -9.065468e-03
#>  [803,]  0.20078716  6.340626e-02 -7.295018e-03
#>  [804,]  0.20818343  7.733752e-02 -2.093172e-03
#>  [805,]  0.19053245  7.512366e-02 -1.488838e-02
#>  [806,]  0.31478714  1.068226e-01 -2.057632e-02
#>  [807,]  0.21730722  4.413645e-02 -1.817738e-02
#>  [808,]  0.22266667  1.101286e-01 -7.750239e-03
#>  [809,]  0.21367546  8.497042e-02 -1.900854e-02
#>  [810,]  0.23470480  1.036726e-01 -1.175097e-02
#>  [811,]  0.17150510  7.160711e-02 -3.444053e-02
#>  [812,]  0.22543095  8.960381e-02 -1.192653e-02
#>  [813,]  0.23021002  9.787798e-02 -1.207585e-02
#>  [814,]  0.21282697  7.268423e-02 -1.132737e-02
#>  [815,]  0.23311085  1.003646e-01 -9.702797e-03
#>  [816,]  0.20591165  8.833905e-02 -5.401715e-03
#>  [817,]  0.19402903  1.376191e-01 -2.876558e-02
#>  [818,]  0.17081511  6.942544e-02  9.965236e-03
#>  [819,]  0.19831849  6.681040e-02 -4.787903e-03
#>  [820,]  0.18886844  7.752408e-02 -1.250679e-02
#>  [821,]  0.16306755  4.377311e-02 -1.034681e-02
#>  [822,]  0.20135840  8.413631e-02 -1.886491e-02
#>  [823,]  0.23741836  9.786412e-02 -1.344432e-02
#>  [824,] -0.21370180  1.152620e-01  5.963890e-02
#>  [825,]  0.21933834  1.098669e-01  2.646216e-02
#>  [826,]  0.20359166  9.606242e-02 -9.013669e-03
#>  [827,]  0.23228202  7.750945e-02 -1.533043e-03
#>  [828,]  0.19902770  9.522858e-02 -2.102949e-02
#>  [829,]  0.23059926  8.763704e-02 -1.095048e-02
#>  [830,]  0.21994646  1.095673e-01 -3.990857e-02
#>  [831,]  0.21239557  8.007209e-02  2.549240e-04
#>  [832,]  0.19457373  7.809750e-02 -3.131459e-03
#>  [833,]  0.20834465  7.820204e-02 -1.341500e-02
#>  [834,]  0.20667500  7.533039e-02 -8.916318e-03
#>  [835,]  0.19683795  7.683206e-02 -2.549117e-02
#>  [836,]  0.20298018  7.480191e-02 -1.400467e-02
#>  [837,]  0.22412481  1.205707e-01  2.810871e-02
#>  [838,]  0.20003128  7.161960e-02 -2.402288e-02
#>  [839,]  0.21217593  3.777244e-02 -2.990574e-02
#>  [840,]  0.21192254  8.330627e-02  1.396939e-03
#>  [841,]  0.22288199  8.766944e-02 -1.684770e-02
#>  [842,]  0.24690036  1.056200e-01 -6.176825e-03
#>  [843,]  0.23297889  9.407601e-02 -1.511875e-02
#>  [844,]  0.21440753  8.339600e-02 -9.925949e-03
#>  [845,]  0.18838070  7.806906e-02 -1.685976e-02
#>  [846,]  0.20297733  6.919753e-02 -1.136693e-02
#>  [847,]  0.24229501  9.390814e-02 -2.250272e-02
#>  [848,]  0.21828461  9.446434e-02 -3.361852e-02
#>  [849,]  0.21297634  8.506853e-02 -1.122066e-02
#>  [850,]  0.22681119  8.018316e-02 -1.409826e-02
#>  [851,]  0.21151754  8.364880e-02 -9.167277e-03
#>  [852,]  0.18402141  5.726392e-02 -1.863072e-02
#>  [853,]  0.21480397  8.689545e-02 -7.716043e-03
#>  [854,]  0.21510065  7.993836e-02 -1.833184e-02
#>  [855,]  0.21089425  8.745580e-02 -1.564040e-02
#>  [856,]  0.20805338  1.143461e-01 -2.040643e-02
#>  [857,]  0.18588362  6.385502e-02 -1.619055e-02
#>  [858,]  0.20526773  6.508191e-02 -1.420178e-02
#>  [859,]  0.22429791  8.993030e-02 -1.340924e-02
#>  [860,]  0.19092962  8.835294e-02 -5.365995e-03
#>  [861,]  0.23581988  9.703587e-02 -7.753562e-03
#>  [862,]  0.17869820  5.562574e-02  5.139294e-02
#>  [863,]  0.22363931  5.701128e-02 -4.116202e-02
#>  [864,]  0.20666104  8.494806e-02 -6.079794e-03
#>  [865,]  0.23010623  9.673695e-02 -1.063618e-02
#>  [866,]  0.21785075  8.488298e-02  7.349744e-03
#>  [867,]  0.20032524  8.943650e-02 -3.358691e-02
#>  [868,]  0.20866095  7.449841e-02 -1.475542e-02
#>  [869,]  0.21866835  1.293540e-01  3.142075e-02
#>  [870,]  0.22427101  9.919368e-02  4.043885e-02
#>  [871,]  0.21882153  7.331200e-02 -1.024617e-02
#>  [872,]  0.29344746 -6.726770e-02 -4.248045e-02
#>  [873,]  0.18298495  7.162502e-02 -9.491923e-03
#>  [874,]  0.21299149  8.594601e-02 -2.326687e-04
#>  [875,]  0.20209893  7.812327e-02 -4.062517e-02
#>  [876,]  0.16524584  5.900805e-02 -3.840909e-02
#>  [877,]  0.17960536  3.192250e-02  8.788147e-04
#>  [878,]  0.20117462  3.462708e-02 -6.301850e-02
#>  [879,]  0.21606335  9.333607e-02 -2.786160e-02
#>  [880,]  0.21225882  1.063604e-01  9.542570e-03
#>  [881,]  0.22289308  9.275161e-02 -7.582511e-03
#>  [882,]  0.19415394  6.892851e-02 -6.322023e-03
#>  [883,]  0.20802290  7.780461e-02 -1.246833e-02
#>  [884,]  0.44527744 -5.000000e-01 -7.047408e-02
#>  [885,]  0.20581303  6.449064e-02 -8.614800e-03
#>  [886,]  0.19775425  1.301243e-01 -9.917769e-02
#>  [887,]  0.19966863  7.016988e-02 -1.310927e-02
#>  [888,]  0.17753474  7.542320e-02 -1.990137e-02
#>  [889,]  0.19141608  7.197119e-02 -1.322652e-02
#>  [890,]  0.28121501  1.162433e-01  5.466009e-03
#>  [891,]  0.18574571  9.157420e-02 -1.338512e-02
#>  [892,]  0.21766544  8.714376e-02 -1.339565e-02
#>  [893,]  0.23414620  9.087376e-02 -8.755708e-03
#>  [894,]  0.19894757  8.514189e-02 -4.697196e-03
#>  [895,]  0.23426042  8.986819e-02 -9.381142e-03
#>  [896,]  0.20778545  7.632117e-02  1.421864e-02
#>  [897,]  0.19490047  1.004939e-01 -2.000495e-02
#>  [898,]  0.24377949  1.048518e-01 -1.147203e-02
#>  [899,]  0.23082422  9.049644e-02 -1.765335e-02
#>  [900,]  0.20491076  7.548560e-02 -1.681817e-02
#>  [901,]  0.24681194  8.369426e-02 -1.883901e-02
#>  [902,]  0.21867242  7.847979e-02 -2.058315e-02
#>  [903,]  0.18896014  5.537334e-02 -1.937249e-02
#>  [904,]  0.21949221  8.648882e-02 -1.497087e-02
#>  [905,]  0.20628445  7.569865e-02 -1.379606e-02
#>  [906,]  0.18959739  5.154054e-02 -1.462575e-02
#>  [907,]  0.18241529  5.314797e-02 -3.244672e-02
#>  [908,]  0.20016710  6.551346e-02 -1.915297e-02
#>  [909,]  0.20941853  7.953915e-02 -1.988155e-03
#>  [910,]  0.19604607  7.049524e-02 -1.392994e-02
#>  [911,]  0.20991888  8.120003e-02  1.682685e-02
#>  [912,]  0.21144646  7.768566e-02 -5.660066e-03
#>  [913,]  0.10552959 -1.264738e-01 -2.163152e-01
#>  [914,]  0.20886729  7.845108e-02 -1.902988e-02
#>  [915,]  0.49170157  1.368088e-01 -1.307057e-01
#>  [916,]  0.21793915  9.645967e-02  2.798388e-02
#>  [917,]  0.20998797  8.135592e-02 -2.202830e-02
#>  [918,]  0.20072840  7.980427e-02 -3.116382e-02
#>  [919,]  0.17974177  6.416616e-02 -2.129086e-02
#>  [920,]  0.21206887  7.871606e-02  2.647413e-02
#>  [921,]  0.18232686  6.632908e-02 -1.351344e-02
#>  [922,]  0.22588089  8.949017e-02 -1.301281e-02
#>  [923,]  0.19511974  6.647923e-02 -2.674605e-02
#>  [924,]  0.18244285  1.164810e-01  3.083187e-03
#>  [925,]  0.21142902  8.419959e-02 -4.705651e-04
#>  [926,]  0.19773479  7.389783e-02 -1.361056e-02
#>  [927,]  0.23955278  8.112846e-02  3.492158e-02
#>  [928,]  0.20450635  7.664535e-02 -1.237598e-02
#>  [929,]  0.17759894  7.771034e-02 -4.288507e-02
#>  [930,]  0.19878135  7.429615e-02 -1.639524e-02
#>  [931,]  0.22021674  7.845713e-02 -7.030190e-03
#>  [932,]  0.22696211  8.800231e-02 -9.304761e-03
#>  [933,]  0.24297241  1.035171e-01 -1.302021e-02
#>  [934,]  0.21334311  8.364973e-02 -1.296120e-02
#>  [935,]  0.23125609  1.396898e-01 -3.937677e-02
#>  [936,]  0.23590644  1.019628e-01 -2.109303e-03
#>  [937,]  0.22190944  9.533983e-02 -7.486705e-03
#>  [938,]  0.23761408  1.076215e-01 -5.277747e-05
#>  [939,]  0.20071641  4.985734e-02  9.997872e-03
#>  [940,]  0.16461801  7.242905e-02 -1.857016e-02
#>  [941,]  0.23596912  2.607167e-02  9.604565e-02
#>  [942,]  0.21203304  8.328755e-02 -1.702006e-02
#>  [943,]  0.23616537  1.001922e-01 -1.314273e-02
#>  [944,]  0.19733390  1.261960e-01 -7.964108e-03
#>  [945,]  0.23397970  9.688603e-02 -1.510607e-02
#>  [946,]  0.16999786  5.678485e-02 -2.121867e-02
#>  [947,]  0.20642007  7.899465e-02 -9.730801e-03
#>  [948,]  0.19370611  7.212202e-02 -1.432317e-02
#>  [949,]  0.25064903  3.541109e-02 -1.230968e-02
#>  [950,]  0.20972995  8.755281e-02 -1.190255e-02
#>  [951,]  0.23397084  1.130290e-01  1.167132e-02
#>  [952,]  0.19880698  5.052281e-02 -2.326324e-02
#>  [953,]  0.23924598  6.774484e-02 -9.781434e-03
#>  [954,]  0.21684758  6.397469e-02 -8.049738e-03
#>  [955,]  0.30257072  1.070446e-01 -2.119736e-02
#>  [956,]  0.06309842  5.000000e-01 -1.192345e-01
#>  [957,]  0.18762772  6.682757e-02 -1.700685e-02
#>  [958,]  0.19490155  6.566159e-02 -1.096495e-02
#>  [959,]  0.19393856  5.309751e-02 -1.326447e-02
#>  [960,]  0.18060322  6.114793e-02  3.640810e-04
#>  [961,]  0.22689827  7.319978e-02  1.886610e-03
#>  [962,]  0.20210030  6.786002e-02 -2.199077e-02
#>  [963,]  0.20008637  6.495250e-02 -1.417164e-02
#>  [964,]  0.21326037  5.265050e-02 -3.157793e-02
#>  [965,]  0.23531812  9.504233e-02 -8.933697e-03
#>  [966,]  0.17784364  8.147498e-02 -1.831436e-02
#>  [967,]  0.21429763  9.235955e-02 -7.317302e-03
#>  [968,]  0.20558026  7.785778e-02 -1.998171e-02
#>  [969,]  0.37026114  7.818104e-02  5.771023e-02
#>  [970,]  0.22214752  8.624716e-02 -1.000715e-02
#>  [971,]  0.22727639  9.054548e-02 -7.917561e-03
#>  [972,]  0.24259259  1.014642e-01 -1.285398e-02
#>  [973,]  0.23300825  9.535473e-02 -1.330015e-02
#>  [974,]  0.23109602  1.055247e-01  8.313882e-03
#>  [975,]  0.23127510  8.411981e-02 -7.045805e-03
#>  [976,]  0.22407323  7.343162e-02 -2.025743e-02
#>  [977,]  0.20207527  7.813550e-02 -1.426891e-02
#>  [978,]  0.23638215  1.005635e-01 -7.130481e-03
#>  [979,]  0.21409508  7.909927e-02 -1.537117e-02
#>  [980,]  0.21000241  7.927070e-02 -1.135459e-02
#>  [981,]  0.20196820  7.577968e-02 -2.179566e-02
#>  [982,]  0.21886054  1.008626e-01  1.145512e-03
#>  [983,]  0.21949257  7.158876e-02 -1.829181e-02
#>  [984,]  0.21170667  8.149323e-02 -1.626514e-02
#>  [985,]  0.15984656  6.979781e-02 -1.706804e-02
#>  [986,]  0.17132781  7.006968e-02 -1.662074e-02
#>  [987,]  0.22514346  1.036527e-01 -1.265193e-02
#>  [988,]  0.21701166  8.870777e-02 -1.586692e-02
#>  [989,]  0.18872924  6.888354e-02 -1.555867e-02
#>  [990,]  0.17985703  7.263466e-02 -1.952350e-02
#>  [991,]  0.24118838  9.730776e-02 -1.186144e-02
#>  [992,]  0.22537723  2.992985e-02 -4.257838e-02
#>  [993,]  0.21816442  8.934273e-02 -7.065846e-03
#>  [994,]  0.19203743  7.047344e-02 -2.611628e-03
#>  [995,]  0.23359763  4.335684e-02 -1.254597e-01
#>  [996,]  0.20484152  5.931845e-02 -1.992099e-02
#>  [997,]  0.19344300  6.102781e-02 -8.819594e-03
#>  [998,]  0.21002239  1.142043e-01  1.803163e-02
#>  [999,]  0.20531905  7.706917e-02 -8.985112e-03
#> [1000,]  0.22879681  9.877686e-02 -1.127042e-02
#> attr(,"addend")
#> [1]  0.03478889 -0.03369141 -0.02612384
#> attr(,"multiplier")
#> [1] 1.425828